% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap.R
\name{all_data}
\alias{all_data}
\title{Get values of all data}
\arguments{
\item{obj}{a \code{\link{taxmap}} object}

\item{...}{Passed to \code{\link{all_names}}}
}
\value{
Named \code{list}
}
\description{
Get values of all data in a \code{\link{taxmap}} object
\preformatted{
obj$data_used(...)
data_used(obj, ...)}
}
\examples{
ex_taxmap$all_data()

}
\seealso{
Other accessors: \code{\link{all_names}},
  \code{\link{data_used}}, \code{\link{get_data}},
  \code{\link{names_used}}
}

