% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--man.R
\name{all_names}
\alias{all_names}
\title{Return names of data in a \code{\link{taxmap}}}
\arguments{
\item{obj}{(\code{\link{taxmap}}) The \code{\link{taxmap}} object containing
taxon information to be queried.}

\item{tables}{If \code{TRUE}, include the names of columns of tables in
\code{obj$data}}

\item{funcs}{If \code{TRUE}, include the names of user-definable functionsin
\code{obj$funcs}.}

\item{others}{If \code{TRUE}, include the names of data in \code{obj$data}
besides tables.}

\item{builtin_funcs}{If \code{TRUE}, include functions like
\code{\link{n_supertaxa}} that provide information for each taxon.}

\item{warn}{If \code{TRUE}, warn if there are duplicate names.}
}
\value{
\code{character}
}
\description{
Return all the valid names that can be used with non-standard evalulation in
manipulation functions like \code{filter_taxa}. \preformatted{
obj$all_names(tables = TRUE, funcs = TRUE, others = TRUE, warn = FALSE)
all_names(obj, tables = TRUE, funcs = TRUE, others = TRUE, warn = FALSE)}
}
\examples{
# Get the names of all data accesible by non-standard evaluation
ex_taxmap$all_names()

# Dont include the names of functions
ex_taxmap$all_names(funcs = FALSE)

}
\seealso{
Other accessors: \code{\link{all_data}},
  \code{\link{data_used}}, \code{\link{get_data}},
  \code{\link{names_used}}
}

