% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id.R
\name{as_tibble.taxa_taxon_id}
\alias{as_tibble.taxa_taxon_id}
\title{#' @export
as.data.frame.taxa_taxon_id <- function(x, row.names = NULL, optional = FALSE, ...,
stringsAsFactors = default.stringsAsFactors()) {
cbind(
data.frame(tax_id = as.character(x), row.names = row.names, stringsAsFactors = stringsAsFactors, ...),
as.data.frame(tax_db(x), row.names = row.names, stringsAsFactors = stringsAsFactors, ...)
)
}}
\usage{
\method{as_tibble}{taxa_taxon_id}(x, ...)
}
\description{
#' @export
as.data.frame.taxa_taxon_id <- function(x, row.names = NULL, optional = FALSE, ...,
stringsAsFactors = default.stringsAsFactors()) {
cbind(
data.frame(tax_id = as.character(x), row.names = row.names, stringsAsFactors = stringsAsFactors, ...),
as.data.frame(tax_db(x), row.names = row.names, stringsAsFactors = stringsAsFactors, ...)
)
}
}
