% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-class.R
\name{binomial}
\alias{binomial}
\title{binomial}
\usage{
binomial(genus, epithet = NULL, canonical = NULL, species = NULL,
  authority = NULL)
}
\arguments{
\item{genus}{genus name}

\item{epithet}{epithet name}

\item{canonical}{canonical name}

\item{species}{genus + epithet}

\item{authority}{taxonomic authority}
}
\description{
binomial
}
\examples{
poa <- binomial("Poa")
poa$tojson()
}

