% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{bison_bison}
\alias{bison_bison}
\title{A data.frame representing a taxonomic classification.}
\description{
Data is from GBIF for the taxon \emph{Bison bison}
}
\details{
It has the following columns

\itemize{
\item name. taxon name
\item rank. taxon rank
\item id. taxon id
}
}
\keyword{data}
