% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{char_or_placeholder}
\alias{char_or_placeholder}
\title{Convert to character or a placeholder}
\usage{
char_or_placeholder(thing, placeholder = "[none]")
}
\arguments{
\item{thing}{The thing to convert to a character}

\item{placeholder}{What is returned when "thing" is NULL or of zero length.}
}
\description{
The placeholder is used with the thing to be printed is NULL or of zero length.
}
\keyword{internal}
