% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_arg_class}
\alias{check_arg_class}
\title{Check argument types}
\usage{
check_arg_class(value, valid_classes, arg_name = deparse(substitute(x)))
}
\arguments{
\item{value}{The value of the argument}

\item{valid_classes}{A character vector of valid classes. It must inherit at least one of these.}

\item{arg_name}{The name of the argument used in error messages. By defualt the name of the
variable passed to "value" is used.}
}
\description{
Check that an argument is one of an accepted set of classes and throw and error if it is not.
}
\keyword{internal}
