% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--internal.R
\name{check_taxmap_data}
\alias{check_taxmap_data}
\title{Check dataset format}
\usage{
check_taxmap_data(obj)
}
\arguments{
\item{obj}{A \code{\link[=taxmap]{taxmap()}} object}
}
\description{
Check that the datasets in a \code{\link[=taxmap]{taxmap()}} object are in the correct format.
\itemize{
\item Checks that column names are not the names of functions
}
}
\keyword{internal}
