% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--column_generators.R
\name{classifications}
\alias{classifications}
\title{Get classification of taxa}
\usage{
classifications(obj, sep = ";")
}
\arguments{
\item{obj}{(\code{\link{taxmap}})}

\item{sep}{(\code{character} of length 1)
The character(s) to place between taxon IDs}
}
\value{
\code{character} of length equal to \code{subset}
}
\description{
Get classification strings of taxa in an object of type \code{\link{taxmap}}.
Each classification is constructed by concatenating the taxon ids of the given taxon and its supertaxa.
}
\examples{
classifications(ex_taxmap)

}
\seealso{
Other taxon_funcs: \code{\link{n_obs_1}},
  \code{\link{n_obs}}, \code{\link{n_subtaxa_1}},
  \code{\link{n_subtaxa}}, \code{\link{n_supertaxa}}
}

