% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{clone_if_r6}
\alias{clone_if_r6}
\title{Clone R6 objects}
\usage{
clone_if_r6(input)
}
\arguments{
\item{input}{the object to clone}
}
\description{
Check if the object is an R6 class and return a deep clone if so, otherwise return the object
}
\keyword{internal}
