% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_database.R
\docType{data}
\name{database_list}
\alias{database_list}
\title{database list}
\format{An object of class \code{list} of length 8.}
\usage{
database_list
}
\description{
database list
}
\keyword{datasets}
