% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{database_definitions}
\alias{database_definitions}
\title{Valid taxonomy databases}
\format{An object of class \code{list} of length 3.}
\usage{
database_definitions
}
\arguments{
\item{name}{(character) name of the database}

\item{url}{(character) url for the database}

\item{desc}{(character) description of the database}

\item{id_regex}{(character) id regex}

\item{rank_levels}{Possible taxonomic ranks}
}
\description{
This defines the valid taxonomic databases that can be used in \code{taxa_database} objects as a list of
}
\section{Attribution}{


This code is based on the code handling options in \link{knitr}.
}

\examples{

# List all database definitions
database_definitions$get()

# Get a specific database definition
database_definitions$get('ncbi')

# Add or overwrite a database definition
database_definitions$set(
  name = "my_new_database",
  url = "http://www.my_tax_database.com",
  desc = "I just made this up",
  id_regex = ".*"
)

# Reset definitions to default values
database_definitions$reset()

}
\keyword{datasets}
