% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{database_ref}
\alias{database_ref}
\title{Database list}
\format{
An object of class \code{taxa_taxon_db_def} (inherits from \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 8.
}
\usage{
database_ref
}
\description{
The list of known databases. Not currently used much, but will be when we add
more check for taxon IDs and taxon ranks from particular databases.
}
\keyword{internal}
