% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank.R
\name{db_ranks}
\alias{db_ranks}
\title{ID validation regexs from database names}
\usage{
db_ranks(db)
}
\arguments{
\item{db}{The names of the databases}
}
\value{
regexes as character vector
}
\description{
ID validation regexs from database names
}
\keyword{internal}
