% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_definitions.R
\name{default_database_definitions}
\alias{default_database_definitions}
\title{Defines valid taxonomic databases}
\usage{
default_database_definitions(defaults = list())
}
\arguments{
\item{name}{(character) name of the database}

\item{url}{(character) url for the database}

\item{description}{(character) description of the database}

\item{id_regex}{(character) id regex}
}
\description{
Defines valid taxonomic databases
}
\section{Attribution}{


This code is copied from the code handling options in \link{knitr}.
}

\keyword{internal}
