% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{delete_unused_class_taxa}
\alias{delete_unused_class_taxa}
\title{Removes taxa from the taxonomy of a classification that are not used by any of the instances}
\usage{
delete_unused_class_taxa(x)
}
\description{
Removes taxa from the taxonomy of a classification that are not used by any of the instances
}
\keyword{internal}
