% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{duplicated_index}
\alias{duplicated_index}
\title{Returns the index of the first occurrence of each unique element}
\usage{
duplicated_index(x, proxy_func = c)
}
\description{
Returns the index of the first occurrence of each unique element
}
\keyword{internal}
