% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{duplicated_index_taxonomy}
\alias{duplicated_index_taxonomy}
\title{Returns the index of the first occurrence of each unique taxon in a taxonomy}
\usage{
duplicated_index_taxonomy(x, proxy_func = NULL)
}
\description{
Returns the index of the first occurrence of each unique taxon in a taxonomy
}
\keyword{internal}
