% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{eval_many}
\alias{eval_many}
\title{Evaluate multiple expressions}
\arguments{
\item{obj}{a \code{\link[=taxonomy]{taxonomy()}} or \code{\link[=taxmap]{taxmap()}} object}

\item{...}{One or more expressions}
}
\value{
A \code{list} of evaluated expressions
}
\description{
Evaluate multiple expressions using
\href{http://adv-r.had.co.nz/Computing-on-the-language.html}{non-standard evaluation}.
Symbols (i.e., variable names) in the expressions will be replaced with the
values associated with terms in \code{\link[=data_used]{data_used()}}, if their names match. Symbols
with the same name in the global envriomnet will be ignored unless they are
prefixed with \code{.env$}.
\preformatted{
obj$eval_many(...)}
}
\examples{
# Evaluating expressions
ex_taxmap$eval_many(n_legs + 1, sum(n_obs))

# Forcing use of variables in the global envrionment
n_legs = 5
ex_taxmap$eval_many(n_legs + 1, .env$n_legs + 1)

}
\seealso{
Other NSE helpers: \code{\link{all_names}},
  \code{\link{data_used}}, \code{\link{eval_one}},
  \code{\link{get_data}}, \code{\link{names_used}}
}
\concept{NSE helpers}
