% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{eval_one}
\alias{eval_one}
\title{Evaluate an expressions}
\arguments{
\item{obj}{A \code{\link[=taxonomy]{taxonomy()}} or \code{\link[=taxmap]{taxmap()}} object}

\item{input}{An expression (i.e. any R code)}
}
\value{
The input expression evaluated
}
\description{
Evaluate an expression using
\href{http://adv-r.had.co.nz/Computing-on-the-language.html}{non-standard evaluation}.
Symbols (i.e., variable names) in the expression will be replaced with the
values associated with terms in \code{\link[=data_used]{data_used()}}, if their names match. Symbols
with the same name in the global envriomnet will be ignored unless they are
prefixed with \code{.env$}.
\preformatted{
obj$eval_one(...)}
}
\examples{
# Evaluating an expression
ex_taxmap$eval_one(sum(n_obs + 1))

# Forcing use of variables in the global envrionment
n_legs = 5
ex_taxmap$eval_many(n_legs + 1)
ex_taxmap$eval_many(.env$n_legs + 1)

}
\seealso{
Other NSE helpers: \code{\link{all_names}},
  \code{\link{data_used}}, \code{\link{eval_many}},
  \code{\link{get_data}}, \code{\link{names_used}}
}
\concept{NSE helpers}
