% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{ex_hierarchy1}
\alias{ex_hierarchy1}
\title{An example Hierarchy object}
\format{A \code{\link[=hierarchy]{hierarchy()}} object.}
\source{
Created from the example code in the \code{\link[=hierarchy]{hierarchy()}}
documentation.
}
\usage{
ex_hierarchy1
}
\description{
An example Hierarchy object built from the ground up.
}
\seealso{
Other taxa-datasets: \code{\link{ex_hierarchies}},
  \code{\link{ex_hierarchy2}}, \code{\link{ex_taxmap}},
  \code{\link{ex_taxonomy}}
}
\keyword{datasets}
