% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{ex_taxmap}
\alias{ex_taxmap}
\title{An example taxmap object}
\format{A \code{\link[=taxmap]{taxmap()}} object.}
\source{
Created from the example code in the \code{\link[=taxmap]{taxmap()}}
documentation.
}
\usage{
ex_taxmap
}
\description{
An example taxmap object built from the ground up. Typically, data stored in
taxmap would be parsed from an input file, but this data set is just for
demonstration purposes.
}
\keyword{datasets}
