% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{ex_taxonomy}
\alias{ex_taxonomy}
\title{An example Taxonomy object}
\format{A \code{\link[=taxonomy]{taxonomy()}} object.}
\source{
Created from the example code in the \code{\link[=taxonomy]{taxonomy()}}
documentation.
}
\usage{
ex_taxonomy
}
\description{
An example Taxonomy object built from the ground up.
}
\keyword{datasets}
