% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--parsers.R
\name{extract_tax_data}
\alias{extract_tax_data}
\title{Extracts taxonomy info from vectors with regex}
\usage{
extract_tax_data(tax_data, key, regex, class_key = "name",
  class_regex = "(.*)", class_sep = NULL, class_rev = FALSE,
  database = "ncbi", return_match = FALSE, return_input = FALSE)
}
\arguments{
\item{tax_data}{A vector from which to extract taxonomy information.}

\item{key}{(\code{character}) The identity of the capturing groups defined using
\code{regex}. The length of \code{key} must be equal to the number of capturing
groups specified in \code{regex}. Any names added to the terms will be used as
column names in the output. Only \code{"info"} can be used multiple times. Each
term must be one of those decribed below:
\itemize{
\item \code{taxon_id}: A unique numeric id for a taxon for a particular \code{database}
(e.g. ncbi accession number). Requires an internet connection.
\item \code{taxon_name}: The name of a taxon. Not necessarily unique, but are
interpretable by a particular \code{database}. Requires an internet connection.
\item \code{class}: A list of taxa information that constitutes the full taxonomic
classification from broad to specific (see \code{class_rev}) for a particular
\code{database}. Individual taxa are separated by the \code{class_sep} argument and
the information is parsed by the \code{class_regex} and \code{class_key} arguments.
\item \code{info}: Arbitrary taxon info you want included in the output. Can be used
more than once.
}}

\item{regex}{(\code{character; length == 1}) A regular expression with capturing
groups indicating the locations of relevant information. The identity of
the information must be specified using the \code{key} argument.}

\item{class_key}{(\code{character} of length 1) The identity of the capturing
groups defined using \code{class_regex}. The length of \code{class_key} must be equal
to the number of capturing groups specified in \code{class_regex}. Any names
added to the terms will be used as column names in the output. At least
\code{"taxon_id"} or \code{"name"} must be specified. Only \code{"taxon_info"} can be used
multiple times. Each term must be one of those decribed below:
\itemize{
\item \code{taxon_id}: A unique numeric id for a taxon for a particular \code{database}
(e.g. ncbi accession number). Requires an internet connection.
\item \code{name}: The name of a taxon. Not necessarily unique, but are
interpretable by a particular \code{database}. Requires an internet connection.
\item \code{info}: Arbitrary taxon info you want included in the output. Can be used
more than once.
}}

\item{class_regex}{(\code{character} of length 1)
A regular expression with capturing groups indicating the locations of data
for each taxon in the \code{class} term in the \code{key} argument. The identity of
the information must be specified using the \code{class_key} argument. The
\code{class_sep} option can be used to split the classification into data for
each taxon before matching. If \code{class_sep} is \code{NULL}, each match of
\code{class_regex} defines a taxon in the classification.}

\item{class_sep}{(\code{character} of length 1)
Used with the \code{class} term in the \code{key} argument. The character(s) used to
separate individual taxa within a classification. After the string defined
by the \code{class} capture group in \code{regex} is split by \code{class_sep}, its
capture groups are extracted by \code{class_regex} and defined by \code{class_key}.
If \code{NULL}, every match of \code{class_regex} is used instead with first
splitting by \code{class_sep}.}

\item{class_rev}{(\code{logical} of length 1)
Used with the \code{class} term in the \code{key} argument. If \code{TRUE}, the order of
taxon data in a classfication is reversed to be specific to broad.}

\item{database}{(\code{character} of length 1) The name of the database that
patterns given in \code{parser} will apply to. Valid databases include "ncbi",
"itis", "eol", "col", "tropicos", "nbn", and "none". \code{"none"} will cause no
database to be quired; use this if you want to not use the internet. NOTE:
Only \code{"ncbi"} has been tested so far.}

\item{return_match}{(\code{logical} of length 1) If \code{TRUE}, include the part of
the input matched by \code{regex} in the output object.}

\item{return_input}{(\code{logical} of length 1) If \code{TRUE}, include the input in
the output object.}
}
\value{
Returns an object of type \code{taxmap}
}
\description{
The location and identity of important information in the input is specified
using a regular expression with capture groups and an corresponding key. An
object of type \code{taxmap} is returned containing the specifed information.
}
