% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_style.R
\name{font_default}
\alias{font_default}
\title{Default font}
\usage{
font_default(text)
}
\arguments{
\item{text}{What to print.}
}
\description{
A wrapper to make changing the formatting of text printed easier.
}
\seealso{
Other printer fonts: \code{\link{desc_font}},
  \code{\link{error_font}}, \code{\link{font_na}},
  \code{\link{font_punct}}, \code{\link{font_secondary}},
  \code{\link{font_tax_name}}, \code{\link{name_font}},
  \code{\link{punc_font}}, \code{\link{tid_font}}
}
\concept{printer fonts}
\keyword{internal}
