% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_style.R
\name{font_na}
\alias{font_na}
\title{Font for NAs in print methods}
\usage{
font_na(text)
}
\arguments{
\item{text}{What to print}
}
\description{
A simple wrapper to make changing the formatting of text printed easier.
This is used for \code{NA}s
}
\seealso{
Other printer fonts: \code{\link{desc_font}},
  \code{\link{error_font}}, \code{\link{font_punct}},
  \code{\link{font_secondary}},
  \code{\link{font_tax_name}}, \code{\link{name_font}},
  \code{\link{punc_font}}, \code{\link{tid_font}}
}
\concept{printer fonts}
\keyword{internal}
