% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_style.R
\name{font_punct}
\alias{font_punct}
\title{Punctuation formatting in print methods}
\usage{
font_punct(text)
}
\arguments{
\item{text}{What to print}
}
\description{
A simple wrapper to make changing the formatting of text printed easier.
This is used for non-data, formatting characters
}
\seealso{
Other printer fonts: \code{\link{font_default}},
  \code{\link{font_na}}, \code{\link{font_secondary}},
  \code{\link{font_tax_name}}
}
\concept{printer fonts}
\keyword{internal}
