% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_style.R
\name{font_secondary}
\alias{font_secondary}
\title{Font for secondary data}
\usage{
font_secondary(text)
}
\arguments{
\item{text}{What to print.}
}
\description{
A wrapper to make changing the formatting of text printed easier.
This is used for print data associated with other data.
}
\seealso{
Other printer fonts: \code{\link{desc_font}},
  \code{\link{error_font}}, \code{\link{name_font}},
  \code{\link{punc_font}}, \code{\link{tid_font}}
}
\concept{printer fonts}
\keyword{internal}
