% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_style.R
\name{font_tax_name}
\alias{font_tax_name}
\title{Taxon name formatting in print methods}
\usage{
font_tax_name(text)
}
\arguments{
\item{text}{What to print}
}
\description{
A simple wrapper to make changing the formatting of text printed easier.
This is used for taxon names.
}
\seealso{
Other printer fonts: \code{\link{font_default}},
  \code{\link{font_na}}, \code{\link{font_punct}},
  \code{\link{font_secondary}}
}
\concept{printer fonts}
\keyword{internal}
