% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--docs.R
\name{get_dataset}
\alias{get_dataset}
\title{Get a data set from a taxmap object}
\arguments{
\item{obj}{A taxmap object}

\item{dataset}{Which data set to use. Can be any of the following:
\describe{
\item{Name}{The name of the data set to use.}
\item{Index}{The index of the data set to use.}
\item{TRUE/FALSE vector}{A TRUE/FALSE vector the same length as the
number of datasets, with exactly one TRUE corresponding to the
selected data set.}
}}
}
\description{
Get a data set from a taxmap object and complain if it does not
exist.
}
\examples{
\dontrun{
# Get data set by name
get_dataset(ex_taxmap, "info")

# Get data set by indeex_taxmap
get_dataset(ex_taxmap, 1)

# Get data set by T/F vector
get_dataset(ex_taxmap, c(T, F, F, F))

}

}
