% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{get_taxon_field}
\alias{get_taxon_field}
\title{Get a value from taxon_name in taxon}
\usage{
get_taxon_field(x, i, simplify = TRUE, only_first = TRUE)
}
\arguments{
\item{simplify}{If TRUE, the return a vector composed of the first item in each taxon. Otherwise,
a list will all data is returned.}
}
\description{
Get a value from taxon_name in taxon
}
\keyword{internal}
