% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{\%in\%}
\alias{\%in\%}
\title{Value matching for taxa package}
\usage{
x \%in\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\description{
A wrapper for the base value matching \%in\% that is used to take into consideration features of the taxa package.
}
