% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-grouping.R
\name{grouping}
\alias{grouping}
\title{Classification, aka "grouping"}
\usage{
grouping(...)
}
\arguments{
\item{...}{key-value pairs of rank and name. See Details.}
}
\value{
An \code{R6Class} object of class \code{Grouping}
}
\description{
Classification, aka "grouping"
}
\details{
allowed keys:
\itemize{
 \item superkingdom
 \item kingdom
 \item subkingdom
 \item infrakingdom
 \item division
 \item phylum
 \item subphylum
 \item subdivision
 \item infradavision
 \item superclass
 \item class
 \item subclass
 \item infraclass
 \item superorder
 \item order
 \item suborder
 \item infraorder
 \item superfamily
 \item family
 \item subfamily
 \item tribe
 \item subtribe
 \item genus
 \item subgenus
 \item section
 \item subsection
 \item species
 \item subspecies
 \item variety
 \item race
 \item subvariety
 \item stirp
 \item morph
 \item form
 \item aberration
 \item subform
 \item unspecified
}
}
\section{Methods}{

\itemize{
 \item print - print method
 \item names_non_null - collect all entries that are non-NULL
 \item as_df - coerce to a tibble/data.frame
}
}
\examples{
grouping(kingdom = "Animalia")

res <- grouping(
  kingdom = "Animalia",
  phylum = "Chordata",
  class = "Mammalia",
  order = "Artiodactyla",
  family = "Bovidae",
  genus = "Bison",
  species = "Bison bison"
)
res
res$names_non_null()
res$as_df()
res$as_df(wide = TRUE)

# data.frame
library("taxize")
x <- classification("Bison bison", db = "gbif")
gg <- grouping(x[[1]])
gg$names_non_null()
gg
}

