% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-hierarchies.R
\name{hierarchies_}
\alias{hierarchies_}
\title{Make a set of many Hierarchy class objects}
\usage{
hierarchies_(...)
}
\arguments{
\item{...}{Any number of object of class \code{Hierarchy}}
}
\value{
An \code{R6Class} object of class \code{Hierarchy}
}
\description{
Make a set of many Hierarchy class objects
}
\examples{
x <- taxon(
  name = taxon_name("Poaceae"),
  rank = taxon_rank("family"),
  id = taxon_id(4479)
)
y <- taxon(
  name = taxon_name("Poa"),
  rank = taxon_rank("genus"),
  id = taxon_id(4544)
)
z <- taxon(
  name = taxon_name("Poa annua"),
  rank = taxon_rank("species"),
  id = taxon_id(93036)
)
hier1 <- hierarchy(z, y, x)

a <- taxon(
  name = taxon_name("Felidae"),
  rank = taxon_rank("family"),
  id = taxon_id(9681)
)
b <- taxon(
  name = taxon_name("Puma"),
  rank = taxon_rank("genus"),
  id = taxon_id(146712)
)
c <- taxon(
  name = taxon_name("Puma concolor"),
  rank = taxon_rank("species"),
  id = taxon_id(9696)
)
hier2 <- hierarchy(c, b, a)

hierarchies_(hier1, hier2)
}

