% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank_level.R
\name{impute_order_na}
\alias{impute_order_na}
\title{Fill in NA values in sequence}
\usage{
impute_order_na(order, inc = 1)
}
\arguments{
\item{order}{An ascending sequences, possibly with NAs}

\item{inc}{The increment size to use for values in NA blocks at the start and end of the sequence.}
}
\description{
Fill in the NA values in a ascending sequence based on nearby non-NA values.
Used to guess the order values for unknown ranks based on the values of known ranks.
}
\keyword{internal}
