% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--internal.R
\name{init_taxmap_data}
\alias{init_taxmap_data}
\title{Convert \code{data} input for Taxamp}
\usage{
init_taxmap_data(self, data, input_ids, assume_equal = TRUE)
}
\arguments{
\item{self}{The newly created \code{\link[=taxmap]{taxmap()}} object}

\item{data}{The \code{data} variable passed to the \code{Taxmap} constructor}

\item{input_ids}{The taxon IDs for the inputs that made the taxonomy}

\item{assume_equal}{If \code{TRUE}, and a data set length is the same as the
\code{input_ids} length, then assume that \code{input_ids} applies to the data set as
well.}
}
\value{
A \code{data} variable with the right format
}
\description{
Make sure \code{data} is in the right format and complain if it is not. Then, add
a \code{taxon_id} column to data with the same length as the input
}
\keyword{internal}
