% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{is_classification}
\alias{is_classification}
\title{Check if is a classification}
\usage{
is_classification(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if an object is the classification class
}
