% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{is_leaf}
\alias{is_leaf}
\title{Check if taxa are leaves}
\usage{
is_leaf(x)
}
\arguments{
\item{x}{An object with taxon supertaxon-subtaxa relationships, such as \code{\link[=taxonomy]{taxonomy()}} objects.}
}
\description{
Check if each taxon is a leaf. A leaf is a taxon with no subtaxa.
subtaxa.
}
