% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{is_taxon}
\alias{is_taxon}
\title{Check if something is a \link{taxon} object}
\usage{
is_taxon(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if an object is of the \link{taxon} class
}
\examples{
x <- taxon(c('A', 'B', 'C'))
is_taxon(x)
is_taxon(1:2)

}
