% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_auth.R
\name{is_taxon_auth}
\alias{is_taxon_auth}
\title{Check if is a taxon authority}
\usage{
is_taxon_auth(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if an object is the taxon authority class
}
