% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_db.R
\name{is_taxon_db}
\alias{is_taxon_db}
\title{Check if something is a \link{taxon_db}}
\usage{
is_taxon_db(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if an object is of the \link{taxon_db} class
}
\examples{
x <- taxon_db(c('ncbi', 'ncbi', 'itis'))
is_taxon_db(x)
is_taxon_db(1:3)

}
