% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_info.R
\name{is_taxon_info}
\alias{is_taxon_info}
\title{Check if is a taxon info object}
\usage{
is_taxon_info(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if an object is the taxon info class
}
