% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank.R
\name{is_taxon_rank}
\alias{is_taxon_rank}
\title{Check if something is a \link{taxon_rank}}
\usage{
is_taxon_rank(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if an object is of the \link{taxon_rank} class
}
\examples{
x <- taxon_rank(c('species', 'species', 'phylum', 'family'))
is_taxon_rank(x)
is_taxon_rank(1:3)

}
