% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank_level.R
\name{is_taxon_rank_level}
\alias{is_taxon_rank_level}
\title{Check if is a taxon id}
\usage{
is_taxon_rank_level(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if an object is the taxon id class
}
