% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{is_taxonomy}
\alias{is_taxonomy}
\title{Check if something is a \link{taxonomy}}
\usage{
is_taxonomy(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if an object is of the \link{taxonomy} class
}
\examples{
x <- taxonomy(c('Carnivora', 'Felidae', 'Panthera', 'Panthera leo',
                'Panthera tigris', 'Ursidae', 'Ursus', 'Ursus arctos'),
              supertaxa = c(NA, 1, 2, 3, 3, 1, 6, 7))
is_taxonomy(x)
is_taxonomy(1:2)

}
