% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_db_def.R
\name{is_valid_regex}
\alias{is_valid_regex}
\title{Check regex validity}
\usage{
is_valid_regex(text)
}
\arguments{
\item{text}{The putative regex to check.}
}
\description{
Check if a regular expression is valid
}
\keyword{internal}
