% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified-data-class.R
\name{item_data}
\alias{item_data}
\title{Return item data from \code{\link{classified}}}
\usage{
item_data(obj, calculated_cols = TRUE)
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{calculated_cols}{(\code{logical} of length 1) If \code{TRUE}, return calculated columns using
\code{\link{classified}$taxon_funcs}.}
}
\value{
\code{data.frame} with rows corresponding to items in input
}
\description{
Return a table of data associated with items of an object of type
\code{\link{classified}}.
}

