% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{known_taxon_rank_levels}
\alias{known_taxon_rank_levels}
\title{All known taxonomic ranks}
\format{An object of class \code{numeric} of length 52.}
\usage{
known_taxon_rank_levels
}
\description{
A list of taxonomic ranks from all databases used copmbined into a single
vector to make it easier to maintain the relative order of ranks when data
from multiple databases are combined.
}
\section{Attribution}{


This list was adapted from a similar one in \link{taxize}.
}

\keyword{internal}
