% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyhelpers.R
\name{lazy-helpers}
\alias{lazy-helpers}
\alias{ranks}
\alias{nms}
\alias{ids}
\title{Lazy}
\usage{
ranks(...)

nms(...)

ids(...)
}
\arguments{
\item{...}{unquoted variables}
}
\description{
Lazy
}
\section{How do these functions work?}{

Each function is a light wrapper around \code{\link[dplyr:vars]{dplyr::vars()}}, to use lazy
evaluation to be able to pass in unquote variables to save some typing,
though you can still quote and get the same result

These are not used by themselves but inside of \code{\link[=pop]{pop()}}, \code{\link[=pick]{pick()}}, \code{\link[=span]{span()}}
}

\section{Symbols}{

\itemize{
\item \code{:}: items between x and y, inclusive
\item \code{::}: items between x and y, exclusive
\item \code{. <}: all items below rank of x
\item \code{. >}: all items above rank of x
}
}

\section{ranks}{

Ranks can be any character string in the set of acceptable rank
names.
}

\section{nms}{

\code{nms} is named to avoid using \code{names} which would collide with the
fxn \code{\link[base:names]{base::names()}} in Base R. Can pass in any character taxonomic names.
}

\section{ids}{

Ids are any alphanumeric taxonomic identifier. Some database providers
use all digits, but some use a combination of digits and characters.
}

\examples{
ranks(order)
ranks(order, genus)
ranks(order:genus)
ranks(. > genus)

nms(Poaceae)
nms(Poaceae, Poa)
nms(Poaceae:Poa)
nms(. < Poaceae)

ids(4544)
ids(4544, 4479)
ids(4544:4479)
ids(. < 4479)
}
