% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{multi_sep_split}
\alias{multi_sep_split}
\title{Like \code{strsplit}, but with multiple separators}
\usage{
multi_sep_split(input, split, ...)
}
\arguments{
\item{input}{A character vector}

\item{split}{One or more seperators to use to split \code{input}}

\item{...}{Passed to \code{\link[base:strsplit]{base::strsplit()}}}
}
\description{
Splits items in a vector by multiple separators.
}
\keyword{internal}
