% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{n_leaves}
\alias{n_leaves}
\title{Number of leaves per taxon}
\usage{
n_leaves(x)
}
\arguments{
\item{x}{An object with taxon supertaxon-subtaxa relationships, such as \code{\link[=taxonomy]{taxonomy()}} objects.}
}
\description{
Get the number of leaves per taxon. A leaf is a taxon with no subtaxa.
}
