% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--docs.R
\name{n_obs_1}
\alias{n_obs_1}
\title{Count observation assigned in \code{\link[=taxmap]{taxmap()}}}
\arguments{
\item{obj}{(\code{\link[=taxmap]{taxmap()}})}

\item{target}{The name of the list/vector/table in \code{obj$data}}
}
\value{
\code{numeric}
}
\description{
Count observations for each taxon in a data set in a \code{\link[=taxmap]{taxmap()}} object. This
includes observations for the specific taxon but NOT the observations of its
subtaxa. "Observations" in this sense are the items (for list/vectors) or
rows (for tables) in a dataset. By default, observations in the first data
set in the \code{\link[=taxmap]{taxmap()}} object is used.  For example, if the data set is a
table, then a value of 3 for a taxon means that their are 3 rows in that
table assigned to that taxon.
\preformatted{
obj$n_obs_1(target)
n_obs_1(obj, target)}
}
\examples{
n_obs_1(ex_taxmap, "info")

}
\seealso{
Other taxmap data functions: \code{\link{n_obs}}
}
