% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--column_generators.R
\name{n_subtaxa_1}
\alias{n_subtaxa_1}
\title{Get number of subtaxa}
\usage{
n_subtaxa_1(obj)
}
\arguments{
\item{obj}{(\code{\link{taxmap}})}
}
\value{
\code{numeric}
}
\description{
Get number of subtaxa for each taxon in an object of type \code{\link{taxmap}}, not including subtaxa of subtaxa etc.
This does not include subtaxa assigned to subtaxa.
}
\examples{
n_subtaxa_1(ex_taxmap)

}
\seealso{
Other taxon_funcs: \code{\link{classifications}},
  \code{\link{n_obs_1}}, \code{\link{n_obs}},
  \code{\link{n_subtaxa}}, \code{\link{n_supertaxa}}
}

