% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-name.R
\name{name}
\alias{name}
\title{Name class}
\usage{
name(name = NULL, genus = NULL, epithet = NULL, canonical = NULL,
  species = NULL, authority = NULL, id = NULL, src = NULL, uri = NULL)
}
\arguments{
\item{name}{(character) any name}

\item{genus}{(character) genus name}

\item{epithet}{(character) epithet name}

\item{canonical}{(character) canonical name}

\item{species}{(character) genus + epithet}

\item{authority}{(character) taxonomic authority}

\item{id}{(character/numeric) a taxonomic ID, such as from a online
taxonomic database, or perhaps one created by the user}

\item{src}{(character) source of the \code{id}, e.g. NCBI}

\item{uri}{(character) a URI for the taxon, should be that for the
source given in \code{src}}
}
\value{
An \code{R6Class} object of class \code{Name}
}
\description{
Name class
}
\section{Methods}{

\itemize{
 \item print - print method
 \item to_json - serialize all data to JSON, useful for writing out
 to disk, etc.
}
}
\examples{
(poa <- name("Poa"))
(undef <- name("undefined"))
(sp1 <- name("species 1"))
(poa_annua <- name(genus = "Poa", epithet = "annua"))
(poa_annua_l <- name(genus = "Poa", epithet = "annua", authority = "L."))
(bison_bison <- name(
   genus = "Bison", epithet = "Bison", id = 2441176,
   src = "gbif", uri = "http://gbif.org/species/2441176"
))
bison_bison$genus
bison_bison$name
bison_bison$id
bison_bison$to_json()
bison_bison$to_json(pretty = TRUE)

# convert to json
poa_annua$to_json()
}

