% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{new_taxon}
\alias{new_taxon}
\title{Minimal taxon constructor}
\usage{
new_taxon(.names = NULL, taxa = list())
}
\arguments{
\item{.names}{The names of the vector.}

\item{taxa}{The taxa to store as list of \link{taxon_name} vectors. Each item in the list represents a
singel taxon, potentially with multiple names.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_name}
}
\description{
Minimal taxon constructor for internal use. Only use when the input is known to be valid
since few validity checks are done.
}
\keyword{internal}
