% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_auth.R
\name{new_taxon_auth}
\alias{new_taxon_auth}
\title{Minimal taxon_auth constructor}
\usage{
new_taxon_auth(auth = character())
}
\arguments{
\item{auth}{Zero or more taxon authorities. Inputs will be transformed to a \code{character} vector.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_auth}
}
\description{
Minimal taxon_auth constructor for internal use. Only use when the input is known to be valid since
few validity checks are done.
}
\keyword{internal}
