% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_authority.R
\name{new_taxon_authority}
\alias{new_taxon_authority}
\title{Minimal taxon_authority constructor}
\usage{
new_taxon_authority(
  .names = NULL,
  author = character(),
  date = character(),
  citation = character()
)
}
\arguments{
\item{.names}{The names of the vector.}

\item{author}{Zero or more author names.}

\item{date}{Zero or more dates.}

\item{citation}{Zero or more literature citations.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_authority}
}
\description{
Minimal taxon_authority constructor for internal use. Only use when the input is known to be valid since
few validity checks are done.
}
\keyword{internal}
