% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_db_def.R
\name{new_taxon_db_def}
\alias{new_taxon_db_def}
\title{Minimal taxon_db_def constructor}
\usage{
new_taxon_db_def(name = character(), url = character(),
  desc = character(), id_regex = character(), ranks = list())
}
\arguments{
\item{name}{Name of the database in lower case. Inputs will be transformed to a \code{character} vector.}

\item{url}{URL of the database website. Inputs will be transformed to a \code{character} vector.}

\item{desc}{Description of the database. Inputs will be transformed to a \code{character} vector.}

\item{id_regex}{A regular expression for taxon IDs of the database. Inputs will be transformed to a \code{character} vector.}

\item{ranks}{Valid taxonomic ranks for the database. Should be a list of \code{factor}s.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_db_def}
}
\description{
Minimal taxon_db_def constructor for internal use. Only use when the input is known to be valid since
few validity checks are done.
}
\keyword{internal}
