% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id.R
\name{new_taxon_id}
\alias{new_taxon_id}
\title{Minimal taxon_id constructor}
\usage{
new_taxon_id(.names = NULL, id = character(), db = taxon_db())
}
\arguments{
\item{id}{Zero or more taxonomic ids. Inputs will be transformed to a \code{character} vector.}

\item{db}{The name(s) of the database(s) associated with the IDs. If not \code{NA} (the
default), the input must consist of names of databases in \link{database_ref}. The length must be
0, 1, or equal to the number of IDs.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_id}
}
\description{
Minimal taxon_id constructor for internal use. Only use when the input is known to be valid since
few validity checks are done.
}
\keyword{internal}
