% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_info.R
\name{new_taxon_info}
\alias{new_taxon_info}
\title{Minimal taxon_info constructor}
\usage{
new_taxon_info(info = list())
}
\arguments{
\item{info}{A list of arbitrary, user-defined attributes associated with each taxon. Each element
in the list, one per taxon, should be a named list of zero or more items with unique names.
Values in this list can be accessed with the \link{taxon_info} function. All elements in the list do
not need to contain the same attributes.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_info}
}
\description{
Minimal taxon_info constructor for internal use. Only use when the input is
known to be valid since few validity checks are done.
}
\keyword{internal}
