% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank_level.R
\name{new_taxon_rank_level}
\alias{new_taxon_rank_level}
\title{Minimal taxon_rank_level constructor}
\usage{
new_taxon_rank_level(level = character(), order = numeric())
}
\arguments{
\item{level}{Zero or more taxonomic rank names. If a named numeric is
applied, the names are used for levels and the numeric values are used
for the order. Inputs will be transformed to a \code{character} vector.}

\item{order}{Integers that determine the relative order of taxonomic levels.
Inputs will be transformed to a \code{integer} vector. \code{NA}s can be used to
indicate that the order is not known.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_rank_level}
}
\description{
Minimal taxon_rank_level constructor for internal use. Only use when the
input is known to be valid since few validity checks are done.
}
\keyword{internal}
