% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{new_taxonomy}
\alias{new_taxonomy}
\title{Minimal taxonomy constructor}
\usage{
new_taxonomy(taxa = taxon(), supertaxa = integer())
}
\arguments{
\item{taxa}{A \code{\link[=taxon]{taxon()}} vector.}

\item{supertaxa}{The indexes of \code{taxa} for each taxon's supertaxon.}
}
\value{
An \code{S3} object of class \code{taxa_taxon}
}
\description{
Minimal taxonomy constructor for internal use. Only use when the input is known to be valid since
few validity checks are done.
}
\keyword{internal}
