% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--docs.R
\name{obs}
\alias{obs}
\title{Get data indexes associated with taxa}
\arguments{
\item{obj}{(\code{\link{taxmap}}) The \code{\link{taxmap}} object containing
taxon information to be queried.}

\item{data}{Either the name of something in \code{obj$data} that has taxon
information or a an external object with taxon information. For tables,
there must be a column named "taxon_id" and lists/vectors must be named by
taxon ID.}

\item{subset}{(\code{character}) Taxon IDs or indexes for which observation
indexes will be returned. Default: All taxa in \code{obj} will be used.}

\item{recursive}{(\code{logical}) If \code{FALSE}, only return the
observation assigned to the specified input taxa, not subtaxa. If
\code{TRUE}, return all the observations of every subtaxa, etc.}

\item{simplify}{(\code{logical}) If \code{TRUE}, then combine all the results
into a single vector of unique observation indexes.}
}
\value{
If \code{simplify = FALSE}, then a list of vectors of observation
  indexes are returned corresponding to the \code{target} argument. If
  \code{simplify = TRUE}, then the observation indexes for all \code{target}
  taxa are returned in a single vector.
}
\description{
Given a \code{\link{taxmap}} object, return the indexes associated with each
taxon in a given table included in that \code{\link{taxmap}} object.
\preformatted{ obj$obs(data, subset = NULL, recursive = TRUE, simplify =
FALSE) obs(obj, data, subset = NULL, recursive = TRUE, simplify = FALSE)}
}
\examples{
# Get indexes of rows corresponding to each taxon
ex_taxmap$obs("info")

# Get only a subset of taxon indexes
ex_taxmap$obs("info", subset = 1:2)

# Get only a subset of taxon IDs
ex_taxmap$obs("info", subset = c("1", "2"))

# Only return indexes of rows assinged to each taxon explicitly
ex_taxmap$obs("info", recursive = FALSE)

# Lump all row indexes in a single vector
ex_taxmap$obs("info", simplify = TRUE)

}
\seealso{
Other taxmap taxonomy functions: \code{\link{supertaxa}}
}

