% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--internal.R
\name{parse_heirarchies_to_taxonomy}
\alias{parse_heirarchies_to_taxonomy}
\title{Infer edge list from hierarchies}
\usage{
parse_heirarchies_to_taxonomy(heirarchies)
}
\arguments{
\item{A}{list of \code{\link[=hierarchy]{hierarchy()}} objects.}
}
\description{
Infer edge list and unique taxa from hierarchies.
}
\keyword{internal}
