% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--class.R
\name{parse_tax_data}
\alias{parse_tax_data}
\title{Convert one or more data sets to taxmap}
\usage{
parse_tax_data(..., class_cols = 1, class_sep = ";", sep_is_regex = FALSE,
  mappings = NULL)
}
\arguments{
\item{...}{One or more tables, lists, or vectors to parse. The \strong{first} must
contain the names of taxa that represent \href{https://en.wikipedia.org/wiki/Taxonomy_(biology)#Classifying_organisms}{taxonomic classifications}.
\strong{Additional inputs} can be any vector/list/table. Valid representations
of classifications include:
\itemize{
\item A list/vector or table with column(s) of taxon names:
Something like \code{"Animalia;Chordata;Mammalia;Primates;Hominidae;Homo"}. What
separator is used can be changed with the \code{class_sep} option. For tables,
the classification can be spread over multiple columns and the separator
will be applied to each column, although each column could just be single
taxon names with no separator. Use the \code{class_cols} option to specify which
columns have taxon names.
\item A list in which each entry is a classificaiton. For example,
\code{list(c("Animalia", "Chordata", "Mammalia", "Primates", "Hominidae", "Homo"), ...)}.
\item A list of data.frames where each represents a classification with one
taxon per row. The column that contains taxon names is specified using the
\code{class_cols} option.
}}

\item{class_cols}{(\code{character} or \code{integer}) The names or indexes of columns
that contain classifications if the first input is a table. If mutliple
columns are specified, they will be combined in the order given.}

\item{class_sep}{(\code{character}) One or more separators that delineate taxon
names in a classification. For example, if one column had \code{"Homo sapiens"}
and another had \code{"Animalia;Chordata;Mammalia;Primates;Hominidae"}, then
\code{class_sep = c(" ", ";")}. All separators are applied to each column so
order does not matter.}

\item{sep_is_regex}{(\code{TRUE}/\code{FALSE}) Whether or not \code{class_sep} should be
used as a \href{https://en.wikipedia.org/wiki/Regular_expression}{regular expression}.}

\item{mappings}{(named \code{character}) This defines how the taxonomic
information in the first input applies to the other inputs. This option
should have one fewer values than \code{...}, with values corresponding to each
input besides the first. The names of values should be a column in the
first input and the values theselves should be the corresponding column in
additional inputs that contains the same ID varaibles. If there are no
shared variables, you can add \code{NA} as a placeholder, but you could just
leave that data out since it is not benifiting from being in the taxmap
object. In addition to column names, the names/values can be one of the
following:
\itemize{
\item \code{"{{index}}"} : This means to use the index of rows/items
\item \code{"{{name}}"}  : This means to use row/item names.
\item \code{"{{value}}"} : This means to use the values in vectors or lists.
Lists will be converted to vectors using \code{\link[=unlist]{unlist()}}.
}}
}
\description{
Parses taxonomic information and assoiacted data and stores it in a
\code{\link[taxa:taxmap]{taxa::taxmap()}} object. \href{https://en.wikipedia.org/wiki/Taxonomy_(biology)#Classifying_organisms}{Taxonomic classifications}
must be present somewhere in the first input.
}
