% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\title{Pop names out}
\usage{
pop(.data, ...)
}
\arguments{
\item{.data}{Input, object of class taxon}

\item{...}{unquoted rank names, e.g., family}
}
\value{
an object of the same class as passed in
}
\description{
That is, drop them
}
\details{
Only supports \code{Hierarchy} objects for now
}
\examples{
(x <- taxon(
  name = taxon_name("Poaceae"),
  rank = taxon_rank("family"),
  id = taxon_id(4479)
))

(y <- taxon(
  name = taxon_name("Poa"),
  rank = taxon_rank("genus"),
  id = taxon_id(4544)
))

(z <- taxon(
  name = taxon_name("Poa annua"),
  rank = taxon_rank("species"),
  id = taxon_id(93036)
))

(res <- hierarchy(z, y, x))

## single taxonomic group
res \%>\% pop(family)
### more than 1 - remake res object above first
# res \%>\% pop(family, genus)
}
