% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{print_tree}
\alias{print_tree}
\title{Print a text tree}
\arguments{
\item{obj}{A \code{taxonomy} or \code{taxmap} object}

\item{value}{What data to return. Default is taxon names. Any result of
\code{\link[=all_names]{all_names()}} can be used, but it usually only makes sense to use data with
one value per taxon, like taxon names.}
}
\description{
Print a text-based tree of a \code{\link[=taxonomy]{taxonomy()}} or \code{\link[=taxmap]{taxmap()}} object.
}
\examples{
print_tree(ex_taxmap)

}
