% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_style.R
\name{print_with_color}
\alias{print_with_color}
\title{Print that works with color}
\usage{
print_with_color(x, original_length = length(x), ...)
}
\arguments{
\item{x}{What to print, typically a character vector}

\item{original_length}{The length of the full vector if only part was given.}

\item{...}{Passed to \code{print}}
}
\description{
The same as the \code{print} function, but can print colored text. Its a bit of a hack, but the only
way I found to replicate the behavior of \code{print} without rewriting the entire \code{print} function.
}
\keyword{internal}
