% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{printed_classification}
\alias{printed_classification}
\title{Prepare classification for printing}
\usage{
printed_classification(x, color = FALSE)
}
\arguments{
\item{color}{Use color?}
}
\value{
character
}
\description{
Prepare classification for printing. Makes color optional.
}
\keyword{internal}
