% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_auth.R
\name{printed_taxon_auth}
\alias{printed_taxon_auth}
\title{Prepare taxon_auth for printing}
\usage{
printed_taxon_auth(x, color = FALSE)
}
\arguments{
\item{color}{Use color?}
}
\value{
character
}
\description{
Prepare taxon_auth for printing. Makes color optional.
}
\keyword{internal}
