% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_authority.R
\name{printed_taxon_authority}
\alias{printed_taxon_authority}
\title{Prepare taxon_authority for printing}
\usage{
printed_taxon_authority(x, color = FALSE)
}
\arguments{
\item{color}{Use color?}
}
\value{
character
}
\description{
Prepare taxon_authority for printing. Makes color optional.
}
\keyword{internal}
