% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id.R
\name{printed_taxon_id}
\alias{printed_taxon_id}
\title{Prepare taxon_id for printing}
\usage{
printed_taxon_id(x, color = FALSE)
}
\arguments{
\item{color}{Use color?}
}
\value{
character
}
\description{
Prepare taxon_id for printing. Makes color optional.
}
\keyword{internal}
