% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_name.R
\name{printed_taxon_name}
\alias{printed_taxon_name}
\title{Prepare taxon for printing}
\usage{
printed_taxon_name(x, color = FALSE)
}
\arguments{
\item{color}{Use color?}
}
\value{
character
}
\description{
Prepare taxon for printing. Makes color optional.
}
\keyword{internal}
