% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank_level.R
\name{printed_taxon_rank_level}
\alias{printed_taxon_rank_level}
\title{Prepare taxon_rank_level for printing}
\usage{
printed_taxon_rank_level(x, color = FALSE)
}
\arguments{
\item{color}{Use color?}
}
\value{
character
}
\description{
Prepare taxon_rank_level for printing. Makes color optional.
}
\keyword{internal}
