% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{printed_taxonomy}
\alias{printed_taxonomy}
\title{Prepare taxonomy for printing}
\usage{
printed_taxonomy(x, color = FALSE)
}
\arguments{
\item{color}{Use color?}
}
\value{
character
}
\description{
Prepare taxonomy for printing. Makes color optional.
}
\keyword{internal}
