% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--printers.R
\name{punc_font}
\alias{punc_font}
\title{Punctuation formatting in print methods}
\usage{
punc_font(text)
}
\arguments{
\item{text}{What to print}
}
\description{
A simple wrapper to make changing the formatting of text printed easier.
This is used for non-data, formatting characters
}
\seealso{
Other printer fonts: \code{\link{desc_font}},
  \code{\link{error_font}}, \code{\link{font_na}},
  \code{\link{font_punct}}, \code{\link{font_secondary}},
  \code{\link{font_tax_name}}, \code{\link{name_font}},
  \code{\link{tid_font}}
}
\concept{printer fonts}
\keyword{internal}
