% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank.R
\name{rank_level_color_funcs}
\alias{rank_level_color_funcs}
\title{Get font color for levels}
\usage{
rank_level_color_funcs(levels)
}
\description{
Make list of crayon style functions to print taxon rank levels in color.
}
\keyword{internal}
