% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{ranks_ref}
\alias{ranks_ref}
\title{Lookup-table for IDs of taxonomic ranks}
\format{An object of class \code{data.frame} with 34 rows and 2 columns.}
\usage{
ranks_ref
}
\description{
Composed of two columns:
\itemize{
\item rankid - the ordered identifer value. lower values mean higher rank
\item ranks - all the rank names that belong to the same level, with
different variants that mean essentially the same thing
}
}
\keyword{data}
