% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{remove_redundant_names}
\alias{remove_redundant_names}
\title{Remove redundant parts of taxon names}
\arguments{
\item{obj}{A \code{taxonomy} or \code{taxmap} object}
}
\value{
A \code{taxonomy} or \code{taxmap} object
}
\description{
Remove the names of parent taxa in the begining of their children's names in a \code{taxonomy} or \code{taxmap} object.
This is useful for removing genus names in species binomials.
\preformatted{
obj$remove_redundant_names()
remove_redundant_names(obj)}
}
