% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{roots}
\alias{roots}
\title{Get root indexes}
\usage{
roots(x)
}
\arguments{
\item{x}{An object with taxon supertaxon-subtaxa relationships, such as \code{\link[=taxonomy]{taxonomy()}} objects.}
}
\description{
Get the indexs for roots.
}
