% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--internal.R
\name{simplify}
\alias{simplify}
\title{List to vector of unique elements}
\usage{
simplify(input)
}
\arguments{
\item{input}{A list of vectors}
}
\description{
Implements the \code{simplify} option in many functions like \code{\link[=supertaxa]{supertaxa()}}.
Returns unique name-value pairs if all vectors are named.
}
\keyword{internal}
