% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{subtaxa_apply}
\alias{subtaxa_apply}
\title{Apply function to subtaxa of each taxon}
\arguments{
\item{obj}{The \code{\link[=taxonomy]{taxonomy()}} or \code{\link[=taxmap]{taxmap()}} object containing taxon
information to be queried.}

\item{func}{(\code{function}) The function to apply.}

\item{subset}{(\code{character}) \code{taxon_ids} or indexes of
\code{taxon_data} for which supertaxa will be returned. Default: All taxa
in \code{obj} will be used.}

\item{recursive}{(\code{logical} or \code{numeric}) If \code{FALSE}, only return the
supertaxa one rank above the target taxa. If \code{TRUE}, return all the
supertaxa of every supertaxa, etc. Positive numbers indicate the number of
recursions (i.e. number of ranks above the target taxon to return). \code{1} is
equivalent to \code{FALSE}. Negative numbers are equivalent to \code{TRUE}.}

\item{simplify}{(\code{logical}) If \code{TRUE}, then combine all the results into a
single vector of unique values.}

\item{include_input}{(\code{logical}) If \code{TRUE}, the input taxa are included in
the output}

\item{value}{What data to give to the function. Any result of
\code{all_names(obj)} can be used, but it usually only makes sense to use data
that has an associated taxon id.}

\item{...}{Extra arguments are passed to the function.}
}
\description{
Apply a function to the subtaxa for each taxon. This is similar
to using \code{\link[=subtaxa]{subtaxa()}} with \code{\link[=lapply]{lapply()}} or \code{\link[=sapply]{sapply()}}.
\preformatted{
obj$subtaxa_apply(func, subset = NULL, recursive = TRUE,
  simplify = FALSE, include_input = FALSE, value = NULL, ...)
subtaxa_apply(obj, func, subset = NULL, recursive = TRUE,
  simplify = FALSE, include_input = FALSE, value = NULL, ...)}
}
\examples{
# Count number of subtaxa in each taxon
subtaxa_apply(ex_taxmap, length)
}
