% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_setter_s3_generics.R, R/taxon_name.R
\name{tax_auth}
\alias{tax_auth}
\alias{tax_auth<-}
\alias{tax_auth.taxa_taxon_name}
\alias{tax_auth<-.taxa_taxon_name}
\title{Set and get taxon authorities}
\usage{
tax_auth(x)

tax_auth(x) <- value

\method{tax_auth}{taxa_taxon_name}(x)

\method{tax_auth}{taxa_taxon_name}(x) <- value
}
\arguments{
\item{x}{An object with taxon authorities.}

\item{value}{The taxon IDs to set. Inputs will be coerced into a \code{\link[=taxon_id]{taxon_id()}} vector.}
}
\description{
Set and get the taxon authorities in objects that have them, such as \code{\link[=taxon_name]{taxon_name()}} objects.
Note that this sets all the authority information, such as author name, date, and citations.
To set or get just one of part of the authorities, use \code{\link[=tax_author]{tax_author()}}, \code{\link[=tax_date]{tax_date()}}, or \code{\link[=tax_cite]{tax_cite()}} instead.
}
