% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_setter_s3_generics.R, R/taxon.R,
%   R/taxon_name.R
\name{tax_cite}
\alias{tax_cite}
\alias{tax_cite<-}
\alias{tax_cite.taxa_taxon}
\alias{tax_cite<-.taxa_taxon}
\alias{tax_cite.taxa_taxon_name}
\alias{tax_cite<-.taxa_taxon_name}
\title{Set and get taxon authority citations}
\usage{
tax_cite(x)

tax_cite(x) <- value

\method{tax_cite}{taxa_taxon}(x)

\method{tax_cite}{taxa_taxon}(x) <- value

\method{tax_cite}{taxa_taxon_name}(x)

\method{tax_cite}{taxa_taxon_name}(x) <- value
}
\arguments{
\item{x}{An object with taxon authority dates.}

\item{value}{The taxon citations to set. Inputs will be coerced into a \code{\link[=taxon_auth]{taxon_auth()}} vector.}
}
\description{
Set and get the taxon authority citations in objects that have them, such as \code{\link[=taxon_auth]{taxon_auth()}} objects.
}
