% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_name.R
\name{tax_id}
\alias{tax_id}
\alias{tax_id.taxa_taxon_name}
\alias{tax_id<-}
\alias{tax_id<-.taxa_taxon_name}
\title{Set and get taxon ids}
\usage{
tax_id(x)

\method{tax_id}{taxa_taxon_name}(x)

tax_id(x) <- value

\method{tax_id}{taxa_taxon_name}(x) <- value
}
\arguments{
\item{x}{An object with taxon ids.}

\item{value}{The taxon ids to set. Inputs will be coerced into a \code{\link[=taxon_id]{taxon_id()}} vector.}
}
\value{
A \code{\link[=taxon_id]{taxon_id()}} vector
}
\description{
Set and get taxon ids in objects that have them, such as \code{\link[=taxon]{taxon()}} objects.
}
