% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa_}
\alias{taxa_}
\title{A class for multiple taxon objects}
\usage{
taxa_(...)
}
\arguments{
\item{...}{Any number of object of class \code{Taxon}}
}
\value{
An \code{R6Class} object of class \code{Taxon}
}
\description{
Stores one or more \code{\link[=taxon]{taxon()}} objects. This is just a thin wrapper for a list
of \code{\link[=taxon]{taxon()}} objects.
}
\examples{
(x <- taxon(
  name = taxon_name("Poa annua"),
  rank = taxon_rank("species"),
  id = taxon_id(93036)
))
taxa_(x, x, x)
}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}}, \code{\link{taxmap}},
  \code{\link{taxon_database}}, \code{\link{taxon_id}},
  \code{\link{taxon_name}}, \code{\link{taxon_rank}},
  \code{\link{taxonomy}}, \code{\link{taxon}}
}
