% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/taxon.R,
%   R/taxon_authority.R, R/taxon_db.R, R/taxon_db_def.R, R/taxon_id.R,
%   R/taxon_name.R, R/taxon_rank.R, R/taxon_rank_level.R, R/taxonomy.R
\name{taxa_printing_funcs}
\alias{taxa_printing_funcs}
\alias{format.taxa_taxon}
\alias{obj_print_data.taxa_taxon}
\alias{obj_print_footer.taxa_taxon}
\alias{vec_ptype_abbr.taxa_taxon}
\alias{vec_ptype_full.taxa_taxon}
\alias{pillar_shaft.taxa_taxon}
\alias{format.taxa_taxon_authority}
\alias{obj_print_data.taxa_taxon_authority}
\alias{vec_ptype_abbr.taxa_taxon_authority}
\alias{vec_ptype_full.taxa_taxon_authority}
\alias{pillar_shaft.taxa_taxon_authority}
\alias{format.taxa_taxon_db}
\alias{vec_ptype_abbr.taxa_taxon_db}
\alias{vec_ptype_full.taxa_taxon_db}
\alias{obj_print_data.taxa_taxon_db_def}
\alias{vec_ptype_full.taxa_taxon_db_def}
\alias{format.taxa_taxon_id}
\alias{obj_print_data.taxa_taxon_id}
\alias{vec_ptype_abbr.taxa_taxon_id}
\alias{vec_ptype_full.taxa_taxon_id}
\alias{pillar_shaft.taxa_taxon_id}
\alias{format.taxa_taxon_name}
\alias{obj_print_data.taxa_taxon_name}
\alias{obj_print_footer.taxa_taxon_name}
\alias{vec_ptype_abbr.taxa_taxon_name}
\alias{vec_ptype_full.taxa_taxon_name}
\alias{pillar_shaft.taxa_taxon_name}
\alias{format.taxa_taxon_rank}
\alias{obj_print_data.taxa_taxon_rank}
\alias{obj_print_footer.taxa_taxon_rank}
\alias{vec_ptype_abbr.taxa_taxon_rank}
\alias{vec_ptype_full.taxa_taxon_rank}
\alias{pillar_shaft.taxa_taxon_rank}
\alias{format.taxa_taxon_rank_level}
\alias{obj_print_data.taxa_taxon_rank_level}
\alias{vec_ptype_abbr.taxa_taxon_rank_level}
\alias{vec_ptype_full.taxa_taxon_rank_level}
\alias{toString.taxa_taxon_rank_level}
\alias{obj_print_data.taxa_taxonomy}
\alias{obj_print_footer.taxa_taxonomy}
\title{taxa printing functions}
\usage{
\method{format}{taxa_taxon}(x, ...)

obj_print_data.taxa_taxon(x)

obj_print_footer.taxa_taxon(x)

vec_ptype_abbr.taxa_taxon(x)

vec_ptype_full.taxa_taxon(x)

\method{pillar_shaft}{taxa_taxon}(x, ...)

\method{format}{taxa_taxon_authority}(x, ...)

obj_print_data.taxa_taxon_authority(x)

vec_ptype_abbr.taxa_taxon_authority(x)

vec_ptype_full.taxa_taxon_authority(x)

\method{pillar_shaft}{taxa_taxon_authority}(x, ...)

\method{format}{taxa_taxon_db}(x, ...)

vec_ptype_abbr.taxa_taxon_db(x)

vec_ptype_full.taxa_taxon_db(x)

obj_print_data.taxa_taxon_db_def(x)

vec_ptype_full.taxa_taxon_db_def(x)

\method{format}{taxa_taxon_id}(x, ...)

obj_print_data.taxa_taxon_id(x)

vec_ptype_abbr.taxa_taxon_id(x)

vec_ptype_full.taxa_taxon_id(x)

\method{pillar_shaft}{taxa_taxon_id}(x, ...)

\method{format}{taxa_taxon_name}(x, ...)

obj_print_data.taxa_taxon_name(x)

obj_print_footer.taxa_taxon_name(x)

vec_ptype_abbr.taxa_taxon_name(x)

vec_ptype_full.taxa_taxon_name(x)

\method{pillar_shaft}{taxa_taxon_name}(x, ...)

\method{format}{taxa_taxon_rank}(x, ...)

obj_print_data.taxa_taxon_rank(x)

obj_print_footer.taxa_taxon_rank(x)

vec_ptype_abbr.taxa_taxon_rank(x)

vec_ptype_full.taxa_taxon_rank(x)

\method{pillar_shaft}{taxa_taxon_rank}(x, ...)

\method{format}{taxa_taxon_rank_level}(x, ...)

obj_print_data.taxa_taxon_rank_level(x)

vec_ptype_abbr.taxa_taxon_rank_level(x)

vec_ptype_full.taxa_taxon_rank_level(x)

\method{toString}{taxa_taxon_rank_level}(x, ...)

obj_print_data.taxa_taxonomy(x)

obj_print_footer.taxa_taxonomy(x)
}
\description{
Functions used internally for printing information in taxon objects. They have to be exported to
work, but they are not intended to be directly used by most users.
}
\keyword{internal}
