% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_authority.R
\name{taxa_taxon_authority-class}
\alias{taxa_taxon_authority-class}
\alias{taxa_taxon_authority}
\title{Taxon authority class}
\description{
Taxon authority class. See \link{taxon_authority} for more information
}
