% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--class.R
\docType{data}
\name{Taxmap}
\alias{Taxmap}
\title{unname' is neede for some reason.. something to look into...}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Taxmap
}
\description{
unname' is neede for some reason.. something to look into...
}
\keyword{datasets}

