% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-class.R
\docType{data}
\name{Taxon}
\alias{Taxon}
\title{taxon}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Taxon
}
\description{
taxon
}
\examples{
#maketax("Poa", Grouping$new(genus = 'Poa'))
}
\keyword{datasets}

