% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_auth.R
\name{taxon_auth}
\alias{taxon_auth}
\title{Taxon authority class}
\usage{
taxon_auth(auth = character())
}
\arguments{
\item{auth}{Zero or more taxon authorities. Inputs will be transformed to a \code{character} vector.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_auth}
}
\description{
Used to store taxon authorities. This is typically used to store taxon IDs in \code{\link[=taxon]{taxon()}} objects.
}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}}, \code{\link{taxa}},
  \code{\link{taxmap}}, \code{\link{taxon_db_def}},
  \code{\link{taxon_db}}, \code{\link{taxon_id}},
  \code{\link{taxon_info}}, \code{\link{taxon_name}},
  \code{\link{taxon_rank_level}}, \code{\link{taxon_rank}},
  \code{\link{taxonomy}}, \code{\link{taxon}}
}
\concept{classes}
