% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_authority.R
\name{taxon_authority}
\alias{taxon_authority}
\alias{names.taxa_taxon_authority}
\alias{names<-.taxa_taxon_authority}
\alias{is.na.taxa_taxon_authority}
\title{Taxon ID class}
\usage{
taxon_authority(author = character(), date = NA, .names = NULL)

\method{names}{taxa_taxon_authority}(x)

\method{names}{taxa_taxon_authority}(x) <- value

\method{is.na}{taxa_taxon_authority}(x)
}
\arguments{
\item{author}{Zero or more author names.}

\item{date}{Zero or more dates.}

\item{.names}{The names of the vector.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_authority}
}
\description{
\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")}
The function used to create \code{taxon_authority} objects
}
\examples{

# Making new objects
x <- taxon_authority(c('A', 'B', 'C'))
x <- taxon_authority(c('Cham. & Schldl.', 'L.'),
                     date = c('1827', '1753'))

# Manipulating objects
as.character(x)
x[2]
x[2] <- 'ABC'
names(x) <- c('a', 'b')
tax_auth_date(x) <- c('2000', '1234')

# Using as columns in tables
tibble::tibble(x = x, y = 1:2)
data.frame(x = x, y = 1:2)

}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}},
  \code{\link{levels<-.taxa_taxon}},
  \code{\link{tax_auth_date<-}}, \code{\link{taxmap}},
  \code{\link{taxon_id}}, \code{\link{taxon_name}},
  \code{\link{taxonomy}}
}
\concept{classes}
