% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified-data-class.R
\name{taxon_data}
\alias{taxon_data}
\title{Return taxon data from \code{\link{classified}}}
\usage{
taxon_data(obj, ...)
}
\arguments{
\item{obj}{(\code{\link{classified}})}

\item{calculated_cols}{(\code{logical} of length 1) If \code{TRUE}, return calculated columns using
\code{\link{classified}$taxon_funcs}.}

\item{subset}{(\code{character}) The names of columns, either user defined or generated using \code{taxon_funcs}.
Default: All columns.}

\item{drop}{(\code{logical} of length 1) If \code{TRUE}, if \code{subset} is a single column
name, then a \code{vector} is returned instead of a \code{data.frame}}
}
\value{
A \code{data.frame} or \code{vector} with rows corresponding to taxa in input
}
\description{
Return a table of data associated with taxa of and object of type
\code{\link{classified}}.
}

