% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_db.R
\name{taxon_db}
\alias{taxon_db}
\title{Taxon database class}
\usage{
taxon_db(db = character())
}
\arguments{
\item{db}{Zero or more taxonomic database names. Should be a name contained in
\code{names(database_definitions)}. Inputs will be transformed to a \code{character} vector.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_db}
}
\description{
Used to store the names of taxon databases defined in \code{database_definitions}.
}
\examples{

}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}}, \code{\link{taxa}},
  \code{\link{taxmap}}, \code{\link{taxon_db_def}},
  \code{\link{taxon_id}}, \code{\link{taxon_name}},
  \code{\link{taxon_rank}}, \code{\link{taxonomy}},
  \code{\link{taxon}}
}
\concept{classes}
