% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{taxon_ids}
\alias{taxon_ids}
\title{Get taxon IDs}
\arguments{
\item{obj}{The \code{taxonomy} or \code{taxmap} object.}
}
\description{
Return the taxon IDs in a \code{\link[=taxmap]{taxmap()}} object.
They are in the order they appear in the edge list.
}
\details{
\preformatted{
obj$taxon_ids()
taxon_ids(obj, ...)}
}
