% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{taxon_indexes}
\alias{taxon_indexes}
\title{Get taxon indexes}
\arguments{
\item{obj}{The \code{taxonomy} or \code{taxmap} object.}
}
\description{
Return the taxon indexes in a \code{\link[=taxmap]{taxmap()}} object.
They are the indexes of the edge list rows.
}
\details{
\preformatted{
obj$taxon_indexes()
taxon_indexes(obj, ...)}
}
\seealso{
Other taxonomy data functions: \code{\link{id_classifications}},
  \code{\link{n_subtaxa_1}}, \code{\link{n_subtaxa}},
  \code{\link{n_supertaxa}},
  \code{\link{name_classifications}},
  \code{\link{taxon_ids}}, \code{\link{taxon_names}}
}
