% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_info.R
\name{taxon_info}
\alias{taxon_info}
\alias{taxon_info.default}
\alias{taxon_info<-}
\title{Taxon info class}
\usage{
taxon_info(...)

\method{taxon_info}{default}(info = list(), ...)

taxon_info(x) <- value
}
\arguments{
\item{...}{Used to pass arguments to methods and allow methods to used additional arguments.}

\item{info}{A list of arbitrary, user-defined attributes associated with each taxon. Each element
in the list, one per taxon, should be a named list of zero or more items with unique names.
Values in this list can be accessed with the \link{taxon_info} function. All elements in the list do
not need to contain the same attributes.}

\item{x}{An object with taxa.}

\item{value}{The taxa to set. Inputs will be coerced into a list of uniquely named list.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_info}
}
\description{
Used to store arbitrary user-defined data assocaited with data. This is typically used to
store taxon info in \code{\link[=taxon]{taxon()}} objects.
}
\keyword{internal}
