% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_name.R
\name{taxon_name}
\alias{taxon_name}
\alias{taxon_name.default}
\alias{taxon_name<-}
\alias{taxon_name<-.taxa_taxon}
\alias{taxon_name.taxa_taxon}
\alias{is.na.taxa_taxon_name}
\title{Taxon name class}
\usage{
taxon_name(...)

\method{taxon_name}{default}(name = character(), db = NA, ...)

taxon_name(x) <- value

\method{taxon_name}{taxa_taxon}(x) <- value

\method{taxon_name}{taxa_taxon}(x, ...)

\method{is.na}{taxa_taxon_name}(x)
}
\arguments{
\item{...}{Used to pass arguments to methods and allow methods to used additional arguments.}

\item{name}{Zero or more taxonomic ids. Inputs will be transformed to a \code{character} vector.}

\item{db}{The name(s) of the database(s) associated with the IDs. If not \code{NA} (the
default), the input must consist of names of databases in \link{database_ref}. The length must be
0, 1, or equal to the number of IDs.}

\item{x}{An object with taxon names.}

\item{value}{The taxon names to set. Inputs will be coerced into a \code{\link[=taxon_name]{taxon_name()}} vector.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_name}
}
\description{
\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")}
Used to store taxon IDs, either arbitrary or from a taxonomy database. This is typically used to
store taxon IDs in \code{\link[=taxon]{taxon()}} objects.
}
\examples{

# Making new objects
x <- taxon_name(c('Homo sapiens', 'Bacillus', 'Ascomycota', 'Ericaceae'))
x <- taxon_name(c('Homo sapiens', 'Bacillus', 'Ascomycota', 'Ericaceae'),
                db = c('ncbi', 'ncbi', 'itis', 'itis'))

# Manipulating objects
as.character(x)
x[2:3]
x[2:3] <- 'ABC'

# Using as columns in tables
tibble::tibble(x = x, y = 1:4)
data.frame(x = x, y = 1:4)

# Trying to use an invalid ID with a specified database causes an error
#taxon_id('NOLETTERS', db = 'ncbi')

}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}}, \code{\link{taxmap}},
  \code{\link{taxon_db}}, \code{\link{taxon_id}},
  \code{\link{taxon_rank}}, \code{\link{taxonomy}},
  \code{\link{taxon}}
}
\concept{classes}
