% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-class.R
\docType{data}
\name{Taxonref}
\alias{Taxonref}
\title{taxonref}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Taxonref
}
\description{
taxonref
}
\examples{
Taxonref$new("genus", "Poa", 54, "http://scottchamberlain.info/")
}
\keyword{datasets}

