% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--printers.R
\name{tid_font}
\alias{tid_font}
\title{Taxon id formatting in print methods}
\usage{
tid_font(text)
}
\arguments{
\item{text}{What to print}
}
\description{
A simple wrapper to make changing the formatting of text printed easier.
}
\seealso{
Other printer fonts: \code{\link{punc_font}}
}
\keyword{internal}
