% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{to_percent}
\alias{to_percent}
\title{Format a proportion as a printed percent}
\usage{
to_percent(prop, digits = 3, ...)
}
\arguments{
\item{prop}{The proportion}

\item{digits}{how many significant digits are to be used for
    numeric and complex \code{x}.  The default, \code{NULL}, uses
    \code{\link{getOption}("digits")}.  This is a suggestion: enough decimal
    places will be used so that the smallest (in magnitude) number has
    this many significant digits, and also to satisfy \code{nsmall}.
    (For the interpretation for complex numbers see \code{\link{signif}}.)}

\item{...}{passed to \code{format}}
}
\value{
character
}
\description{
Format a proportion as a printed percent
}
\keyword{internal}
