% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--docs.R
\name{transmute_obs}
\alias{transmute_obs}
\title{Replace columns in \code{\link[=taxmap]{taxmap()}} objects}
\arguments{
\item{obj}{An object of type \code{\link[=taxmap]{taxmap()}}}

\item{target}{The name of the table in \code{obj$data} to filter}

\item{...}{One or more named columns to add. Newly created columns can be
referenced in the same function call. Any variable name that appears in
\code{\link[=all_names]{all_names()}} can be used as if it was a vector on its own.}
}
\value{
An object of type \code{\link[=taxmap]{taxmap()}}
}
\description{
Replace columns of tables in \code{obj$data} in \code{\link[=taxmap]{taxmap()}} objects. See
\code{\link[dplyr:transmute]{dplyr::transmute()}} for the inspiration for this function and more
information.
\preformatted{
obj$transmute_obs(target, ...)
transmute_obs(obj, target, ...)}
}
\examples{
transmute_obs(ex_taxmap, "info", new_col = paste0(name, "!!!"))

}
\seealso{
Other taxmap manipulation functions: \code{\link{arrange_obs}},
  \code{\link{arrange_taxa}}, \code{\link{filter_obs}},
  \code{\link{filter_taxa}}, \code{\link{mutate_obs}},
  \code{\link{sample_frac_obs}},
  \code{\link{sample_frac_taxa}},
  \code{\link{sample_n_obs}}, \code{\link{sample_n_taxa}},
  \code{\link{select_obs}}
}
