% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{unname_fields}
\alias{unname_fields}
\title{Remove names from fields in a vctrs rcrd}
\usage{
unname_fields(x)
}
\arguments{
\item{x}{a vctrs rcrd}
}
\description{
Remove names from fields in a vctrs rcrd
}
\keyword{internal}
