% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap.R
\name{arrange_obs}
\alias{arrange_obs}
\title{Sort columns of \code{\link{taxmap}} objects}
\arguments{
\item{obj}{An object of type \code{\link{taxmap}}}

\item{target}{The name of the table in \code{obj$data} to filter}

\item{...}{One or more column names to sort on.}
}
\value{
An object of type \code{\link{taxmap}}
}
\description{
Sort columns of tables in \code{obj$data} in \code{\link{taxmap}} objects.
Any variable name that appears in \code{obj$all_names()} can be used as if it was a vector on its own.
See \code{\link[dplyr]{arrange}} for the inspiration for this function and more information.
\preformatted{
obj$arrange_obs(target, ...)
arrange_obs(obj, target, ...)}
}
\examples{
# Sort in ascending order
arrange_obs(ex_taxmap, "info", n_legs)

# Sort in decending order
arrange_obs(ex_taxmap, "info", desc(n_legs))

}
\seealso{
Other dplyr-like functions: \code{\link{arrange_taxa}},
  \code{\link{filter_obs}}, \code{\link{filter_taxa}},
  \code{\link{mutate_obs}}, \code{\link{sample_frac_obs}},
  \code{\link{sample_frac_taxa}},
  \code{\link{sample_n_obs}}, \code{\link{sample_n_taxa}},
  \code{\link{select_obs}}, \code{\link{transmute_obs}}
}

