% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--docs.R
\name{arrange_taxa}
\alias{arrange_taxa}
\title{Sort the edge list of \code{\link[=taxmap]{taxmap()}} objects}
\arguments{
\item{obj}{\code{\link[=taxmap]{taxmap()}}}

\item{...}{One or more column names to sort on.}
}
\value{
An object of type \code{\link[=taxmap]{taxmap()}}
}
\description{
Sort the edge list in \code{\link[=taxmap]{taxmap()}} objects. Any variable name that
appears in \code{obj$all_names()} can be used as if it was a vector on its
own. See \code{\link[dplyr:arrange]{dplyr::arrange()}} for the inspiration for this function
and more information. \preformatted{ obj$arrange_taxa(...) arrange_taxa(obj,
...)}
}
\examples{
arrange_taxa(ex_taxmap, desc(ex_taxmap$taxon_names()))

}
\seealso{
Other dplyr-like functions: \code{\link{arrange_obs}},
  \code{\link{filter_obs}}, \code{\link{filter_taxa}},
  \code{\link{mutate_obs}}, \code{\link{sample_frac_obs}},
  \code{\link{sample_frac_taxa}},
  \code{\link{sample_n_obs}}, \code{\link{sample_n_taxa}},
  \code{\link{select_obs}}, \code{\link{transmute_obs}}
}
