% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{as_taxon}
\alias{as_taxon}
\title{Convert to a taxon vector}
\usage{
as_taxon(x, ...)
}
\arguments{
\item{x}{An object to be converted to a taxon vector}

\item{...}{Additional parameters.}
}
\description{
Convert other objects to taxon vectors
}
