% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Taxon class}
\usage{
classification(
  taxa = taxon(),
  supertaxa = NA,
  instances = integer(),
  .names = NULL
)
}
\arguments{
\item{taxa}{A \code{\link[=taxon]{taxon()}} vector or something that can be converted to a
\code{\link[=taxon]{taxon()}} vector.}

\item{supertaxa}{The indexes of \code{taxa} for each taxon's supertaxon.}

\item{instances}{The indexes of each instance of a taxon in the taxonomy. Can
be any length, but must consist of valid indexes for taxa in the \code{taxonomy}
object.}

\item{.names}{The names of the vector.}
}
\value{
An \code{S3} object of class \code{taxa_classification}
}
\description{
\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")} Used to
store classifications in reference to a taxonomic tree. For more information
on what each class is designed for, see the \link{concepts} section of the help
pages.
}
\examples{

# Create taxon vector
x <- classification(c('Carnivora', 'Felidae', 'Panthera', 'Panthera leo',
                     'Panthera tigris', 'Ursidae', 'Ursus', 'Ursus arctos'),
                    supertaxa = c(NA, 1, 2, 3, 3, 1, 6, 7),
                    instances = c(3, 4, 4, 5, 5, 6, 8, 8, 2, 5, 6, 2))

x <- classification(taxon(name = c('Carnivora', 'Felidae', 'Panthera', 'Panthera leo',
                                   'Panthera tigris', 'Ursidae', 'Ursus', 'Ursus arctos'),
                          rank = c('order', 'family', 'genus', 'species',
                                   'species', 'family', 'genus', 'species'),
                          id = taxon_id(c('33554', '9681', '9688', '9689',
                                          '9694', '9632', '9639', '9644'),
                                        db = 'ncbi'),
                          auth = c('Bowdich, 1821', 'Fischer, 1817', 'Oken, 1816', 'L., 1758',
                                   'L., 1758', 'Fischer, 1817', 'L., 1758', 'L., 1758')),
                    supertaxa = c(NA, 1, 2, 3, 3, 1, 6, 7),
                    instances = c(3, 4, 4, 5, 5, 6, 8, 8, 2, 5, 6, 2))

# Get parts of the classification vector
tax_name(x)
tax_rank(x)
tax_id(x)
tax_db(x)
tax_auth(x)
tax_author(x)
tax_date(x)
tax_cite(x)

# Set parts of the classification vector
tax_name(x) <- tolower(tax_name(x))
tax_rank(x)[1] <- NA
tax_id(x) <- '9999'
tax_db(x) <- 'itis'
tax_auth(x) <- NA
tax_author(x)[2:3] <- c('Joe', 'Billy')
tax_date(x) <- c('1999', '2013', '1796', '1899')
tax_cite(x)[1] <- 'Linnaeus, C. (1771). Mantissa plantarum altera generum.'

# Manipulate classification vectors
x[1:3]
x[tax_rank(x) > 'family']
c(x, x)
names(x) <- c('a', 'b', 'c', 'd')
x['b'] <- NA
is.na(x)
as.data.frame(x)
as_tibble(x)

# Use as columns in tables
tibble::tibble(x = x, y = 1:4)
data.frame(x = x, y = 1:4)

}
\seealso{
Other classes: 
\code{\link{taxon_authority}()},
\code{\link{taxon_db}()},
\code{\link{taxon_id}()},
\code{\link{taxon_name}()},
\code{\link{taxon_rank}()},
\code{\link{taxonomy}()},
\code{\link{taxon}()}
}
\concept{classes}
