% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{data_used}
\alias{data_used}
\title{Get values of data used in expressions}
\arguments{
\item{obj}{a \code{\link[=taxonomy]{taxonomy()}} or \code{\link[=taxmap]{taxmap()}} object}

\item{...}{One or more expressions}
}
\value{
\code{list}
}
\description{
Get values of data in a \code{\link[=taxonomy]{taxonomy()}} or \code{\link[=taxmap]{taxmap()}} used in expressions.
Expressions are not evaluated and do not need to make sense.
\preformatted{
obj$data_used(...)
data_used(obj, ...)}
}
\examples{
data_used(ex_taxmap, n_legs + dangerous == invalid_expression)

}
\seealso{
Other accessors: \code{\link{all_names}},
  \code{\link{get_data}}, \code{\link{names_used}}
}
