% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{db_ref}
\alias{db_ref}
\title{Valid taxonomy databases}
\format{
An object of class \code{list} of length 3.
}
\usage{
db_ref
}
\description{
This defines the valid taxonomic databases that can be used in \code{taxa_database} objects as a list of
}
\section{Attribution}{


This code is based on the code handling options in \link{knitr}.
}

\examples{

# List all database definitions
db_ref$get()

# Get a specific database definition
db_ref$get('ncbi')

# Add or overwrite a database definition
db_ref$set(
  name = "my_new_database",
  url = "http://www.my_tax_database.com",
  desc = "I just made this up",
  id_regex = ".*"
)

# Reset definitions to default values
db_ref$reset()

}
\keyword{datasets}
