% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_style.R
\name{font_secondary}
\alias{font_secondary}
\title{Font for secondary data}
\usage{
font_secondary(text)
}
\arguments{
\item{text}{What to print.}
}
\description{
A wrapper to make changing the formatting of text printed easier.
This is used for print data associated with other data.
}
\seealso{
Other printer fonts: 
\code{\link{font_default}()},
\code{\link{font_na}()},
\code{\link{font_punct}()},
\code{\link{font_tax_name}()}
}
\concept{printer fonts}
\keyword{internal}
