% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--man.R
\name{get_data}
\alias{get_data}
\title{Get data in a taxmap object by name}
\arguments{
\item{obj}{A \code{\link{taxmap}}  object}

\item{name}{(\code{character}) Names of data to return.}
}
\value{
\code{list}
}
\description{
Given a vector of names, return a list of data contained in a
\code{\link{taxmap}} object. \preformatted{
obj$get_data(name)
get_data(obj, name)}
}
\examples{
ex_taxmap$get_data("reaction")

}
\seealso{
Other accessors: \code{\link{all_data}},
  \code{\link{all_names}}, \code{\link{data_used}},
  \code{\link{names_used}}
}

