% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-class.R
\docType{data}
\name{Grouping}
\alias{Grouping}
\title{Classification, or grouping}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Grouping
}
\description{
Classification, or grouping
}
\examples{
Grouping$new()
}
\keyword{datasets}

