% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{internodes}
\alias{internodes}
\title{Get internodes}
\usage{
internodes(x, ...)
}
\arguments{
\item{x}{The object to get internodes for.}

\item{value}{Something to return instead of indexes. Must be the same length as the number of taxa.}
}
\description{
Get internodes indexes for each taxon or another per-taxon value.
}
