% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{is_internode}
\alias{is_internode}
\title{Check if taxa are internodes}
\usage{
is_internode(x)
}
\arguments{
\item{x}{The object to get internodes for.}
}
\description{
Check if each taxon is a internode. An internode is a taxon with exactly one supertaxon and one
subtaxon.
}
