% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{is_root}
\alias{is_root}
\title{Test if taxa are roots}
\usage{
is_root(x)
}
\arguments{
\item{x}{An object with taxon supertaxon-subtaxa relationships, such as \code{\link[=taxonomy]{taxonomy()}} objects.}
}
\description{
Check if each taxon is a root. A root is a taxon with no supertaxon.
}
