% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{leaves}
\alias{leaves}
\title{Get leaves}
\usage{
leaves(x, value = NULL, ...)
}
\arguments{
\item{x}{The object to get leaves for.}

\item{value}{Something to return instead of indexes. Must be the same length as the number of taxa.}
}
\description{
Get leaves indexes for each taxon or another per-taxon value.
}
