% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{leaves}
\alias{leaves}
\title{Get leaf taxa}
\arguments{
\item{obj}{The \code{taxonomy} or \code{taxmap} object containing taxon
information to be queried.}

\item{subset}{(\code{character}) Taxon IDs for which leaf taxa will be
returned. Default: All taxon in \code{obj} will be used.}

\item{return_type}{(\code{logical}) Controls output type: "index", "id",
"taxa", or "hierarchies". Note that "index" is the index of the edge list,
not the taxon list.}

\item{...}{Used by the S3 method to pass the parameters to the R6 method of
\code{\link[=taxonomy]{taxonomy()}}}
}
\value{
\code{character}
}
\description{
Return the leaf taxa for a \code{\link[=taxmap]{taxmap()}} object. Leaf taxa are taxa
with no subtaxa. \preformatted{ obj$leaves(subset = NULL, return_type = "id")
leaves(obj, ...)}
}
