% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--class.R
\name{lookup_tax_data}
\alias{lookup_tax_data}
\title{Convert one or more data sets to taxmap}
\usage{
lookup_tax_data(tax_data, columns, type, datasets, mappings, database,
  include_tax_data)
}
\arguments{
\item{tax_data}{A table, list, or vector that contain sequence IDs, taxon
IDs, or taxon names.
\itemize{
\item tables: The \code{columns} option must be used to specify which column
contains the sequence IDs, taxon IDs, or taxon names.
\item lists: There must be only one item per list entry unless the \code{columns}
option is used to specify what item to use in each list entry.
\item vectors: simply a vector of sequence IDs, taxon IDs, or taxon names.
}}

\item{columns}{(\code{character} or \code{integer}) The names or indexes of columns
that contain classifications if the first input is a table. If mutliple
columns are specified, they will be combined in the order given.}

\item{type}{(\code{"seq_id"}, \code{"taxon_id"}, \code{"taxon_name"}) What type of
information can be used to look up the classifications.}

\item{datasets}{Additional lists/vectors/tables that should be included in
the resulting \code{taxmap} object. The \code{mappings} option is use to specify how
these data sets relate to the \code{tax_data} and, by inference, what taxa apply
to each item.}

\item{mappings}{(named \code{character}) This defines how the taxonomic
information in \code{tax_data} applies to data set in \code{datasets}. This option
should have the same number of inputs as \code{datasets}, with values
corresponding to each data set. The names of the character vector specify
what information in \code{tax_data} is shared with info in each \code{dataset}, which
is specified by the corresponding values of the character vector. If there
are no shared variables, you can add \code{NA} as a placeholder, but you could
just leave that data out since it is not benifiting from being in the
taxmap object. The names/values can be one of the following:
\itemize{
\item For tables, the names of columns can be used.
\item \code{"{{index}}"} : This means to use the index of rows/items
\item \code{"{{name}}"}  : This means to use row/item names.
\item \code{"{{value}}"} : This means to use the values in vectors or lists. Lists
will be converted to vectors using \code{\link[=unlist]{unlist()}}.
}}

\item{database}{(\code{character}) The name of a database to use to look up
classifications.}

\item{include_tax_data}{(\code{TRUE}/\code{FALSE}) Whether or not to include \code{tax_data}
as a dataset, like those in \code{datasets}.}
}
\description{
Looks up taxonomic data from NCBI sequence IDs, taxon IDs, or taxon names
that are present in a dataset. Also can incorperate additional assocaited
datasets.
}
