% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--man.R
\name{n_obs_1}
\alias{n_obs_1}
\title{Count observation assigned in \code{\link{taxmap}}
Count observations assigned to a specific taxon in an \code{\link{taxmap}}.
This does not include observations assigned to subtaxa.
\preformatted{
obj$n_obs_1(target)
n_obs_1(obj, target)}}
\arguments{
\item{obj}{(\code{\link{taxmap}})}

\item{target}{The name of the list/vector/table in \code{obj$data}}
}
\value{
\code{numeric}
}
\description{
Count observation assigned in \code{\link{taxmap}}
Count observations assigned to a specific taxon in an \code{\link{taxmap}}.
This does not include observations assigned to subtaxa.
\preformatted{
obj$n_obs_1(target)
n_obs_1(obj, target)}
}
\examples{
#' n_obs_1(ex_taxmap, "info")

}
\seealso{
Other taxon_funcs: \code{\link{id_classifications}},
  \code{\link{n_obs}}, \code{\link{n_subtaxa_1}},
  \code{\link{n_subtaxa}}, \code{\link{n_supertaxa}},
  \code{\link{name_classifications}}
}

