% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{n_subtaxa}
\alias{n_subtaxa}
\title{Number of subtaxa per taxon}
\usage{
n_subtaxa(x)
}
\arguments{
\item{x}{An object with taxon supertaxon-subtaxa relationships, such as \code{\link[=taxonomy]{taxonomy()}} objects.}
}
\description{
Get the number of subtaxa per taxon.
}
