% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{n_subtaxa}
\alias{n_subtaxa}
\title{Get number of subtaxa}
\arguments{
\item{obj}{(\code{\link{taxmap}})}
}
\value{
\code{numeric}
}
\description{
Get number of subtaxa for each taxon in an object of type \code{\link{taxmap}}
\preformatted{
obj$n_subtaxa()
n_subtaxa(obj)}
}
\examples{
n_subtaxa(ex_taxmap)

}
\seealso{
Other taxon_funcs: \code{\link{id_classifications}},
  \code{\link{n_obs_1}}, \code{\link{n_obs}},
  \code{\link{n_subtaxa_1}}, \code{\link{n_supertaxa}},
  \code{\link{name_classifications}}
}

