% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{n_subtaxa_1}
\alias{n_subtaxa_1}
\title{Get number of subtaxa}
\arguments{
\item{obj}{(\code{\link[=taxmap]{taxmap()}})}
}
\value{
\code{numeric}
}
\description{
Get number of subtaxa for each taxon in an object of type
\code{\link[=taxmap]{taxmap()}}, not including subtaxa of subtaxa etc. This does not
include subtaxa assigned to subtaxa. \preformatted{ obj$n_subtaxa_1()
n_subtaxa_1(obj)}
}
\examples{
n_subtaxa_1(ex_taxmap)

}
\seealso{
Other taxon_funcs: \code{\link{id_classifications}},
  \code{\link{n_obs_1}}, \code{\link{n_obs}},
  \code{\link{n_subtaxa}}, \code{\link{n_supertaxa}},
  \code{\link{name_classifications}}
}
