% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{n_supertaxa}
\alias{n_supertaxa}
\title{Number of supertaxa per taxon}
\usage{
n_supertaxa(x)
}
\arguments{
\item{x}{An object with taxon supertaxon-subtaxa relationships, such as \code{\link[=taxonomy]{taxonomy()}} objects.}
}
\description{
Get the number of supertaxa per taxon.
}
