% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{n_supertaxa}
\alias{n_supertaxa}
\title{Get number of supertaxa}
\arguments{
\item{obj}{(\code{\link{taxmap}})}
}
\value{
\code{numeric}
}
\description{
Get number of supertaxa for each taxon in an object of type
\code{\link{taxmap}} \preformatted{ obj$n_supertaxa() n_supertaxa(obj)}
}
\examples{
n_supertaxa(ex_taxmap)

}
\seealso{
Other taxon_funcs: \code{\link{id_classifications}},
  \code{\link{n_obs_1}}, \code{\link{n_obs}},
  \code{\link{n_subtaxa_1}}, \code{\link{n_subtaxa}},
  \code{\link{name_classifications}}
}

