% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--man.R
\name{names_used}
\alias{names_used}
\title{Get names of data used in expressions}
\arguments{
\item{obj}{a \code{\link{taxmap}} object}

\item{...}{One or more expressions}
}
\value{
Named \code{character}
}
\description{
Get names of available data used in expressions.
Expressions are not evaluated and do not need to make sense.
\preformatted{
obj$names_used(...)
names_used(obj,...)}
}
\examples{
ex_taxmap$names_used(n_legs + dangerous == invalid_expression)

}
\seealso{
Other accessors: \code{\link{all_data}},
  \code{\link{all_names}}, \code{\link{data_used}},
  \code{\link{get_data}}
}

