% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{new_taxon}
\alias{new_taxon}
\title{Minimal taxon constructor}
\usage{
new_taxon(name = taxon_name(), rank = taxon_rank(), id = taxon_id(),
  auth = taxon_auth())
}
\arguments{
\item{name}{The names of taxa. Inputs with be coerced into a \link{taxon_name} vector if anything else
is given.}

\item{rank}{The ranks of taxa. Inputs with be coerced into a \link{taxon_rank} vector if anything else
is given.}

\item{id}{The ids of taxa. These should be unique identifier and are usually associated with a
database. Inputs with be coerced into a \link{taxon_id} vector if anything else is given.}

\item{auth}{The authority of the taxon. Inputs with be coerced into a \link{taxon_auth} vector if
anything else is given.}
}
\value{
An \code{S3} object of class \code{taxa_taxon}
}
\description{
Minimal taxon constructor for internal use. Only use when the input is known to be valid since
few validity checks are done.
}
\keyword{internal}
