% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\title{Pop taxa out}
\usage{
pop(.data, ...)
}
\arguments{
\item{.data}{Input, object of class \link{Hierarchy}, or \link{hierarchies}}

\item{...}{unquoted rank names (e.g., family), taxon names (e.g.,
Poa annua), or taxnomic IDs (e.g., 93036)}
}
\value{
an object of the same class as passed in
}
\description{
That is, drop them
}
\details{
supports \code{Hierarchy} and \code{hierarchies} objects
}
\examples{
ex_hierarchy1
# ranks
pop(ex_hierarchy1, ranks(family))
ex_hierarchy1 \%>\% pop(ranks(family))
ex_hierarchy1 \%>\% pop(ranks(family, genus))
# taxon names
ex_hierarchy1 \%>\% pop(nms(Poa))
ex_hierarchy1 \%>\% pop(nms(Poaceae, Poa))
# taxon ids
ex_hierarchy1 \%>\% pop(ids(4479))
ex_hierarchy1 \%>\% pop(ids(4479, 4544))
# mixed: ids and names
ex_hierarchy1 \%>\% pop(ranks(family), ids(4544))


## single taxonomic group
ex_hierarchy1 \%>\% pop(family)
### more than 1 - remake res object above first
ex_hierarchy1 \%>\% pop(family, genus)

# hierarchies
# single taxonomic group
invisible(lapply(ex_hierarchies, print))
ex_hierarchies \%>\% pop(family) \%>\% lapply(., print) \%>\% invisible

## more than one taxonomic group
invisible(lapply(ex_hierarchies, print))
ex_hierarchies \%>\% pop(family, genus) \%>\% lapply(., print) \%>\% invisible

# Taxonomy class
# ex_taxonomy$pop(ranks(family), NULL, NULL)
# pop(ex_taxonomy, ranks(family))
# pop(ex_taxonomy, ranks(family, genus))
# taxon names
# pop(ex_taxonomy, nms(Poa))
# pop(ex_taxonomy, nms(Poaceae, Poa))
# taxon ids
# pop(ex_taxonomy, ids(4479))
# pop(ex_taxonomy, ids(4479, 4544))
# mixed: ids and names
# pop(ex_taxonomy, ranks(family), ids(4544))
}
