% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank.R
\name{printed_taxon_rank}
\alias{printed_taxon_rank}
\title{Prepare taxon_rank for printing}
\usage{
printed_taxon_rank(x, color = FALSE)
}
\arguments{
\item{color}{Use color?}
}
\value{
character
}
\description{
Prepare taxon_rank for printing. Makes color optional.
}
\keyword{internal}
