% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{roots}
\alias{roots}
\title{Get root indexes}
\usage{
roots(x, subset = NULL)
}
\arguments{
\item{x}{An object with taxon supertaxon-subtaxa relationships, such as \code{\link[=taxonomy]{taxonomy()}} objects.}

\item{subset}{The subset of the tree to search for roots to that subset.}
}
\description{
Get the indexs for roots.
}
