% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-taxonomy.R
\name{roots}
\alias{roots}
\title{Get root taxa}
\arguments{
\item{obj}{The \code{taxonomy} or \code{taxmap} object containing taxon information to be
queried.}

\item{subset}{(\code{character}) Taxon IDs for which supertaxa will be returned. Default: All
taxon in \code{obj} will be used.}

\item{index}{(\code{logical}) If \code{TRUE}, return the indexes of roots in
\code{taxon_data} instead of \code{taxon_ids}}

\item{...}{Used by the S3 method to pass the parameters to the R6 method of \code{\link{taxonomy}}}
}
\value{
\code{character}
}
\description{
Return the root taxa for a \code{\link{taxmap}} object. Can also be used to get the roots of
a subset of taxa.
\preformatted{
obj$roots(subset = NULL, index = FALSE)
roots(obj, ...)}
}

