% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{stems}
\alias{stems}
\title{Get stems}
\usage{
stems(x, value = NULL, ...)
}
\arguments{
\item{x}{The object to get stems for.}

\item{value}{Something to return instead of indexes. Must be the same length as the number of taxa.}

\item{...}{Additional arguments.}
}
\description{
Get stem indexes for each taxon or another per-taxon value.
}
