% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{stems}
\alias{stems}
\title{Get stem taxa}
\arguments{
\item{obj}{The \code{taxonomy} or \code{taxmap} object containing taxon
information to be queried.}

\item{subset}{(\code{character}) Taxon IDs for which stem taxa will be
returned. Default: All taxon in \code{obj} will be used.}

\item{return_type}{(\code{logical}) Controls output type: "index", "id",
"taxa", or "hierarchies". Note that "index" is the index of the edge list,
not the taxon list.}

\item{simplify}{(\code{logical}) If \code{TRUE}, then combine all the results
into a single vector of unique values.}

\item{exclude_leaves}{(\code{logical}) If \code{TRUE}, the do not include
taxa with no subtaxa.}

\item{...}{Used by the S3 method to pass the parameters to the R6 method of
\code{\link[=taxonomy]{taxonomy()}}}
}
\value{
\code{character}
}
\description{
Return the stem taxa for a \code{\link[=taxonomy]{taxonomy()}} or a \code{\link[=taxmap]{taxmap()}}
object. Stem taxa are all those from the roots to the first taxon with more
than one subtaxon. \preformatted{ obj$stems(subset = NULL, simplify = FALSE,
return_type = "id", exclude_leaves = FALSE) stems(obj, ...)}
}
\seealso{
Other taxonomy indexing functions: \code{\link{leaves}},
  \code{\link{roots}}, \code{\link{subtaxa}},
  \code{\link{supertaxa}}
}
