% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{subtaxa}
\alias{subtaxa}
\title{Get subtaxa}
\arguments{
\item{obj}{The \code{taxonomy} or \code{taxmap} object containing taxon information to be
queried.}

\item{subset}{(\code{character}) \code{taxon_ids} or indexes of \code{taxon_data} for which
supertaxa will be returned. Default: All taxa in \code{obj} will be used.}

\item{recursive}{(\code{logical}) If \code{FALSE}, only return the subtaxa one level bwlow the
target taxa. If \code{TRUE}, return all the subtaxa of every subtaxa, etc.}

\item{simplify}{(\code{logical}) If \code{TRUE}, then combine all the results into a single
vector of unique values.}

\item{include_input}{(\code{logical}) If \code{TRUE}, the input taxa are included in the output}

\item{return_type}{(\code{logical}) Controls output type: "index", "id", "taxa", or "hierarchies".
Note that "index" is the index of the edge list, not the taxon list.}
}
\value{
If \code{simplify = FALSE}, then a list of vectors are returned corresponding to the
  \code{target} argument. If \code{simplify = TRUE}, then the unique values are returned in a
  single vector.
}
\description{
Return the taxon IDs or \code{taxon_data} indexes of all subtaxa in an object of type \code{taxmap}
\preformatted{
obj$subtaxa(subset = NULL, recursive = TRUE,
            simplify = FALSE, include_input = FALSE,
            return_type = "id", na = FALSE)
subtaxa(obj, ...)}
}

