% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{supertaxa}
\alias{supertaxa}
\title{Get all supertaxa of a taxon}
\arguments{
\item{obj}{The \code{taxonomy} or \code{taxmap} object containing taxon
information to be queried.}

\item{subset}{(\code{character}) \code{taxon_ids} or indexes of
\code{taxon_data} for which supertaxa will be returned. Default: All taxa
in \code{obj} will be used.}

\item{recursive}{(\code{logical} or \code{numeric}) If \code{FALSE}, only return the
supertaxa one rank above the target taxa. If \code{TRUE}, return all the
supertaxa of every supertaxa, etc. Positive numbers indicate the number of
recursions (i.e. number of ranks above the target taxon to return). \code{1} is
equivalent to \code{FALSE}. Negative numbers are equivalent to \code{TRUE}.}

\item{simplify}{(\code{logical}) If \code{TRUE}, then combine all the results into a
single vector of unique values.}

\item{include_input}{(\code{logical}) If \code{TRUE}, the input taxa are included in
the output}

\item{value}{What data to return. This is usually the name of column in a
table in \code{obj$data}. Any result of \code{all_names(obj)} can be used, but it
usually only makes sense to data that corresponds to taxa 1:1, such as
\code{\link[=taxon_ranks]{taxon_ranks()}}. By default, taxon indexes are returned.}

\item{na}{(\code{logical}) If \code{TRUE}, return \code{NA} where information
is not available.}

\item{...}{U Used by the S3 method to pass the parameters to the R6 method of
\code{\link[=taxonomy]{taxonomy()}}}
}
\value{
If \code{simplify = FALSE}, then a list of vectors are returned
corresponding to the \code{subset} argument. If \code{simplify = TRUE},
then unique values are returned in a single vector.
}
\description{
Return the taxon IDs or indexes of all supertaxa (i.e. all taxa the target
taxa are a part of) in an object of type \code{\link[=taxonomy]{taxonomy()}} or
\code{\link[=taxmap]{taxmap()}}. \preformatted{ obj$supertaxa(subset = NULL, recursive =
TRUE, simplify = FALSE, include_input = FALSE, value = NULL, na =
FALSE) supertaxa(obj, ...)}
}
\seealso{
Other taxonomy indexing functions: \code{\link{leaves}},
  \code{\link{roots}}, \code{\link{stems}},
  \code{\link{subtaxa}}
}
