% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_name.R
\name{tax_auth}
\alias{tax_auth}
\alias{tax_auth.taxa_taxon_name}
\alias{tax_auth<-}
\alias{tax_auth<-.taxa_taxon_name}
\title{Set and get taxon authorities}
\usage{
tax_auth(x)

\method{tax_auth}{taxa_taxon_name}(x)

tax_auth(x) <- value

\method{tax_auth}{taxa_taxon_name}(x) <- value
}
\arguments{
\item{x}{An object with taxon authorities.}

\item{value}{The taxon authorities to set. Inputs will be coerced into a \code{\link[=taxon_auth]{taxon_auth()}} vector.}
}
\value{
A \code{\link[=character]{character()}} vector
}
\description{
Set and get taxon authorities in objects that have them, such as \code{\link[=taxon]{taxon()}} objects.
}
