% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_authority.R
\name{tax_author}
\alias{tax_author}
\alias{tax_author.taxa_taxon_authority}
\title{Set and get taxon authors}
\usage{
tax_author(x)

\method{tax_author}{taxa_taxon_authority}(date = character())
}
\arguments{
\item{x}{An object with taxon authors.}
}
\value{
A \code{\link[=character]{character()}} vector
}
\description{
Set and get taxon authors in objects that have them, such as \code{\link[=taxon]{taxon()}} objects.
}
