% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R,
%   R/getter_setter_s3_generics.R, R/taxon.R, R/taxon_name.R, R/taxonomy.R
\name{tax_id.taxa_classification}
\alias{tax_id.taxa_classification}
\alias{tax_id<-.taxa_classification}
\alias{tax_id}
\alias{tax_id<-}
\alias{tax_id.taxa_taxon}
\alias{tax_id<-.taxa_taxon}
\alias{tax_id.taxa_taxon_name}
\alias{tax_id<-.taxa_taxon_name}
\alias{tax_id.taxa_taxonomy}
\alias{tax_id<-.taxa_taxonomy}
\title{Set and get taxon IDs}
\usage{
\method{tax_id}{taxa_classification}(x)

\method{tax_id}{taxa_classification}(x) <- value

tax_id(x)

tax_id(x) <- value

\method{tax_id}{taxa_taxon}(x)

\method{tax_id}{taxa_taxon}(x) <- value

\method{tax_id}{taxa_taxon_name}(x)

\method{tax_id}{taxa_taxon_name}(x) <- value

\method{tax_id}{taxa_taxonomy}(x)

\method{tax_id}{taxa_taxonomy}(x) <- value
}
\arguments{
\item{x}{An object with taxon IDs.}

\item{value}{The taxon IDs to set. Inputs will be coerced into a \code{\link[=taxon_id]{taxon_id()}} vector.}
}
\description{
Set and get the taxon IDs in objects that have them, such as \code{\link[=taxon_name]{taxon_name()}} objects.
}
