% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_name.R
\name{tax_rank}
\alias{tax_rank}
\alias{tax_rank.taxa_taxon_name}
\alias{tax_rank<-}
\alias{tax_rank<-.taxa_taxon_name}
\title{Set and get taxon ranks}
\usage{
tax_rank(x)

\method{tax_rank}{taxa_taxon_name}(x)

tax_rank(x) <- value

\method{tax_rank}{taxa_taxon_name}(x) <- value
}
\arguments{
\item{x}{An object with taxon ranks.}

\item{value}{The taxon ranks to set. Inputs will be coerced into a \code{\link[=taxon_rank]{taxon_rank()}} vector.}
}
\value{
A \code{\link[=taxon_rank]{taxon_rank()}} vector
}
\description{
Set and get taxon ranks in objects that have them, such as \code{\link[=taxon]{taxon()}} objects.
}
