% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_setter_s3_generics.R, R/taxon.R,
%   R/taxon_name.R, R/taxonomy.R
\name{tax_rank}
\alias{tax_rank}
\alias{tax_rank<-}
\alias{tax_rank.taxa_taxon}
\alias{tax_rank<-.taxa_taxon}
\alias{tax_rank.taxa_taxon_name}
\alias{tax_rank<-.taxa_taxon_name}
\alias{tax_rank.taxa_taxonomy}
\alias{tax_rank<-.taxa_taxonomy}
\title{Set and get taxon ranks}
\usage{
tax_rank(x)

tax_rank(x) <- value

\method{tax_rank}{taxa_taxon}(x)

\method{tax_rank}{taxa_taxon}(x) <- value

\method{tax_rank}{taxa_taxon_name}(x)

\method{tax_rank}{taxa_taxon_name}(x) <- value

\method{tax_rank}{taxa_taxonomy}(x)

\method{tax_rank}{taxa_taxonomy}(x) <- value
}
\arguments{
\item{x}{An object with taxon ranks.}

\item{value}{The taxon ranks to set. Inputs will be coerced into a \code{\link[=taxon_rank]{taxon_rank()}} vector.}
}
\description{
Set and get the taxon ranks in objects that have them, such as \code{\link[=taxon_name]{taxon_name()}} objects.
}
