% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa_}
\alias{taxa_}
\title{Make a set of many Taxon class objects}
\usage{
taxa_(...)
}
\arguments{
\item{...}{Any number of object of class \code{Taxon}}
}
\value{
An \code{R6Class} object of class \code{Taxon}
}
\description{
Make a set of many Taxon class objects
}
\examples{
(x <- taxon(
  name = taxon_name("Poa annua"),
  rank = taxon_rank("species"),
  id = taxon_id(93036)
))
taxa_(x, x, x)
}

