% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/taxon.R, R/taxon_db.R,
%   R/taxon_db_def.R, R/taxon_id.R, R/taxon_info.R, R/taxon_name.R,
%   R/taxon_rank.R, R/taxon_rank_level.R
\name{taxa_casting_funcs}
\alias{taxa_casting_funcs}
\alias{vec_cast.taxa_taxon}
\alias{vec_cast.taxa_taxon.default}
\alias{vec_cast.taxa_taxon.taxa_taxon}
\alias{vec_cast.taxa_taxon.character}
\alias{vec_cast.character.taxa_taxon}
\alias{vec_cast.taxa_taxon.factor}
\alias{vec_cast.factor.taxa_taxon}
\alias{vec_cast.taxa_taxon_db}
\alias{vec_cast.taxa_taxon_db.default}
\alias{vec_cast.taxa_taxon_db.taxa_taxon_db}
\alias{vec_cast.taxa_taxon_db.character}
\alias{vec_cast.character.taxa_taxon_db}
\alias{vec_cast.taxa_taxon_db.factor}
\alias{vec_cast.factor.taxa_taxon_db}
\alias{vec_cast.taxa_taxon_db_def}
\alias{vec_cast.taxa_taxon_db_def.default}
\alias{vec_cast.taxa_taxon_db_def.taxa_taxon_db_def}
\alias{vec_cast.taxa_taxon_id}
\alias{vec_cast.taxa_taxon_id.default}
\alias{vec_cast.taxa_taxon_id.taxa_taxon_id}
\alias{vec_cast.taxa_taxon_id.character}
\alias{vec_cast.character.taxa_taxon_id}
\alias{vec_cast.taxa_taxon_id.factor}
\alias{vec_cast.factor.taxa_taxon_id}
\alias{vec_cast.taxa_taxon_id.integer}
\alias{vec_cast.integer.taxa_taxon_id}
\alias{vec_cast.data.frame.taxa_taxon_id}
\alias{vec_cast.taxa_taxon_info}
\alias{vec_cast.taxa_taxon_info.default}
\alias{vec_cast.taxa_taxon_info.taxa_taxon_info}
\alias{vec_cast.taxa_taxon_info.list}
\alias{vec_cast.list.taxa_taxon_info}
\alias{vec_cast.taxa_taxon_name}
\alias{vec_cast.taxa_taxon_name.default}
\alias{vec_cast.taxa_taxon_name.taxa_taxon_name}
\alias{vec_cast.taxa_taxon_name.character}
\alias{vec_cast.character.taxa_taxon_name}
\alias{vec_cast.taxa_taxon_name.factor}
\alias{vec_cast.factor.taxa_taxon_name}
\alias{vec_cast.taxa_taxon_name.double}
\alias{vec_cast.double.taxa_taxon_name}
\alias{vec_cast.data.frame.taxa_taxon_name}
\alias{vec_cast.taxa_taxon_rank}
\alias{vec_cast.taxa_taxon_rank.default}
\alias{vec_cast.taxa_taxon_rank.taxa_taxon_rank}
\alias{vec_cast.taxa_taxon_rank.character}
\alias{vec_cast.character.taxa_taxon_rank}
\alias{vec_cast.taxa_taxon_rank.factor}
\alias{vec_cast.factor.taxa_taxon_rank}
\alias{vec_cast.taxa_taxon_rank.double}
\alias{vec_cast.double.taxa_taxon_rank}
\alias{vec_cast.data.frame.taxa_taxon_rank}
\alias{vec_cast.taxa_taxon_rank_level}
\alias{vec_cast.taxa_taxon_rank_level.default}
\alias{vec_cast.taxa_taxon_rank_level.taxa_taxon_rank_level}
\alias{vec_cast.taxa_taxon_rank_level.character}
\alias{vec_cast.character.taxa_taxon_rank_level}
\title{taxa casting functions}
\usage{
\method{vec_cast}{taxa_taxon}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon}{default}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon}{taxa_taxon}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon}{character}(x, to, x_arg, to_arg)

\method{vec_cast.character}{taxa_taxon}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon}{factor}(x, to, x_arg, to_arg)

\method{vec_cast.factor}{taxa_taxon}(x, to, x_arg, to_arg)

\method{vec_cast}{taxa_taxon_db}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_db}{default}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_db}{taxa_taxon_db}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_db}{character}(x, to, x_arg, to_arg)

\method{vec_cast.character}{taxa_taxon_db}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_db}{factor}(x, to, x_arg, to_arg)

\method{vec_cast.factor}{taxa_taxon_db}(x, to, x_arg, to_arg)

\method{vec_cast}{taxa_taxon_db_def}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_db_def}{default}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_db_def}{taxa_taxon_db_def}(x, to, x_arg,
  to_arg)

\method{vec_cast}{taxa_taxon_id}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_id}{default}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_id}{taxa_taxon_id}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_id}{character}(x, to, x_arg, to_arg)

\method{vec_cast.character}{taxa_taxon_id}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_id}{factor}(x, to, x_arg, to_arg)

\method{vec_cast.factor}{taxa_taxon_id}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_id}{integer}(x, to, x_arg, to_arg)

\method{vec_cast.integer}{taxa_taxon_id}(x, to, x_arg, to_arg)

\method{vec_cast.data.frame}{taxa_taxon_id}(x, to, x_arg, to_arg)

\method{vec_cast}{taxa_taxon_info}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_info}{default}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_info}{taxa_taxon_info}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_info}{list}(x, to, x_arg, to_arg)

\method{vec_cast.list}{taxa_taxon_info}(x, to, x_arg, to_arg)

\method{vec_cast}{taxa_taxon_name}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_name}{default}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_name}{taxa_taxon_name}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_name}{character}(x, to, x_arg, to_arg)

\method{vec_cast.character}{taxa_taxon_name}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_name}{factor}(x, to, x_arg, to_arg)

\method{vec_cast.factor}{taxa_taxon_name}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_name}{double}(x, to, x_arg, to_arg)

\method{vec_cast.double}{taxa_taxon_name}(x, to, x_arg, to_arg)

\method{vec_cast.data.frame}{taxa_taxon_name}(x, to, x_arg, to_arg)

\method{vec_cast}{taxa_taxon_rank}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_rank}{default}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_rank}{taxa_taxon_rank}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_rank}{character}(x, to, x_arg, to_arg)

\method{vec_cast.character}{taxa_taxon_rank}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_rank}{factor}(x, to, x_arg, to_arg)

\method{vec_cast.factor}{taxa_taxon_rank}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_rank}{double}(x, to, x_arg, to_arg)

\method{vec_cast.double}{taxa_taxon_rank}(x, to, x_arg, to_arg)

\method{vec_cast.data.frame}{taxa_taxon_rank}(x, to, x_arg, to_arg)

\method{vec_cast}{taxa_taxon_rank_level}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_rank_level}{default}(x, to, x_arg, to_arg)

\method{vec_cast.taxa_taxon_rank_level}{taxa_taxon_rank_level}(x, to,
  x_arg, to_arg)

\method{vec_cast.taxa_taxon_rank_level}{character}(x, to, x_arg, to_arg)

\method{vec_cast.character}{taxa_taxon_rank_level}(x, to, x_arg, to_arg)
}
\description{
Functions used internally for casting taxon objects to other types. They have to be exported to
work, but they are not intended to be directly used by most users.
}
\keyword{internal}
