% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/taxon.R, R/taxon_id.R,
%   R/taxon_rank.R, R/taxonomy.R
\name{taxa_comparison_funcs}
\alias{taxa_comparison_funcs}
\alias{vec_proxy_equal.taxa_taxon}
\alias{vec_proxy_equal.taxa_taxon_id}
\alias{vec_proxy_compare.taxa_taxon_rank}
\alias{vec_proxy_equal.taxa_taxon_rank}
\alias{Ops.taxa_taxon_rank}
\alias{vec_proxy_equal.taxa_taxonomy}
\title{taxa comparison functions}
\usage{
\method{vec_proxy_equal}{taxa_taxon}(x, ...)

\method{vec_proxy_equal}{taxa_taxon_id}(x, ...)

\method{vec_proxy_compare}{taxa_taxon_rank}(x, ...)

\method{vec_proxy_equal}{taxa_taxon_rank}(x, ...)

\method{Ops}{taxa_taxon_rank}(e1, e2)

\method{vec_proxy_equal}{taxa_taxonomy}(x, ...)
}
\description{
Functions used internally for casting taxon objects to other types. They have to be exported to
work, but they are not intended to be directly used by most users.
}
\keyword{internal}
