% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-class.R
\name{taxon}
\alias{taxon}
\alias{taxons}
\title{taxon}
\usage{
taxon(name, grouping)

taxons(...)
}
\arguments{
\item{name}{An object of class \code{Name}. Required.}

\item{grouping}{An object of class \code{Grouping}. Required.}

\item{...}{One or more objects of class \code{Taxon}}
}
\value{
An \code{R6Class} object of class \code{Taxon} for function
\code{taxon}, and an object of class \code{Taxa} for function \code{taxa}
}
\description{
taxon
}
\section{Methods}{

\itemize{
 \item print - print method, both \code{taxon} and \code{taxa}
 \item as_df - coerce all groupings to a tibble/data.frame, only \code{taxa}
}
}
\examples{
# a taxon
nm1 <- name("Bison bison")
gr1 <- grouping(
  kingdom = "Animalia",
  family = "Bovidae",
  species = "Bison bison"
)
(t1 <- taxon(nm1, gr1))

# another taxon
nm2 <- name("Poa annua")
gr2 <- grouping(
  kingdom = "Plantae",
  family = "Poaceae",
  species = "Poa annua"
)
(t2 <- taxon(nm2, gr2))

# a taxa object
t3 <- taxons(t1, t2)
t3$taxa
t3$taxa[[1]]
t3$taxa[[2]]
t3$as_df()
t3$taxa[[2]]$name
t3$taxa[[2]]$grouping
t3$taxa[[2]]$grouping$as_df()
}

