% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{taxon}
\alias{taxon}
\title{Taxon class}
\usage{
taxon(name, rank = NULL, id = NULL, authority = NULL)
}
\arguments{
\item{name}{(character) an TaxonName object or character. if character
passed in, we'll coerce to a TaxonName object internally, required}

\item{rank}{(character) an TaxonRank object. if character
passed in, we'll coerce to a TaxonRank object internally, required}

\item{id}{(character) an TaxonId object. if character
passed in, we'll coerce to a TaxonId object internally, required}

\item{authority}{(character) an TaxonAuthority object, optional}
}
\value{
An \code{R6Class} object of class \code{Taxon}
}
\description{
Taxon class
}
\section{Methods}{

\itemize{
\item print - print method
}
}

\examples{
(x <- taxon(
  name = taxon_name("Poa annua"),
  rank = taxon_rank("species"),
  id = taxon_id(93036)
))
x$name
x$id
x$database
}
