% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{taxon_auth}
\alias{taxon_auth}
\alias{taxon_auth.taxa_taxon}
\alias{taxon_auth<-}
\alias{taxon_auth<-.taxa_taxon}
\title{Set and get taxon authorities}
\usage{
taxon_auth(x)

\method{taxon_auth}{taxa_taxon}(x)

taxon_auth(x) <- value

\method{taxon_auth}{taxa_taxon}(x) <- value
}
\arguments{
\item{x}{An object with taxon authorities.}

\item{value}{The taxon authorities to set. Inputs will be coerced into a \code{\link[=character]{character()}} vector.}
}
\value{
A \code{\link[=character]{character()}} vector
}
\description{
Set and get taxon authorities in objects that have them, such as \code{\link[=taxon]{taxon()}} objects.
}
