% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_authority.R, R/taxon_name.R
\name{taxon_authority}
\alias{taxon_authority}
\alias{names.taxa_taxon_authority}
\alias{names<-.taxa_taxon_authority}
\alias{is.na.taxa_taxon_authority}
\alias{names.taxa_taxon_name}
\alias{names<-.taxa_taxon_name}
\title{Taxon authority class}
\usage{
taxon_authority(author = character(), date = NA, citation = NA,
  .names = NA, extract_date = TRUE)

\method{names}{taxa_taxon_authority}(x)

\method{names}{taxa_taxon_authority}(x) <- value

\method{is.na}{taxa_taxon_authority}(x)

\method{names}{taxa_taxon_name}(x)

\method{names}{taxa_taxon_name}(x) <- value
}
\arguments{
\item{author}{Zero or more author names.}

\item{date}{Zero or more dates.}

\item{citation}{Zero or more literature citations.}

\item{.names}{The names of the vector.}

\item{extract_date}{If \code{TRUE} (the default), then if a date is detected in the \code{author} input and
no \code{date} input is given, then the date is seperated from the author input.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_authority}
}
\description{
\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")} The function used to create
\code{taxon_authority} objects
}
\examples{

# Making new objects
x <- taxon_authority(c('A', 'B', 'C'))
x <- taxon_authority(c('Cham. & Schldl.', 'L.'),
                     date = c('1827', '1753'))

# Manipulating objects
as.character(x)
as.data.frame(x)
as_tibble(x)
x[2]
x[2] <- 'ABC'
names(x) <- c('a', 'b')
tax_author(x)[1] <- tolower(tax_author(x)[1])
tax_author(x)
tax_date(x) <- c('2000', '1234')
tax_date(x)
tax_cite(x)[2] <- c('Linnaeus, C. (1771). Mantissa plantarum altera generum.')
tax_cite(x)

# Using as columns in tables
tibble::tibble(x = x, y = 1:2)
data.frame(x = x, y = 1:2)

}
\seealso{
Other classes: \code{\link{classification}},
  \code{\link{taxon_db}}, \code{\link{taxon_id}},
  \code{\link{taxon_name}}, \code{\link{taxon_rank}},
  \code{\link{taxonomy}}, \code{\link{taxon}}
}
\concept{classes}
