% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_db.R, R/taxon_id.R, R/taxon_name.R,
%   R/taxon_rank.R
\name{taxon_db}
\alias{taxon_db}
\alias{taxon_db.default}
\alias{taxon_db<-}
\alias{is.na.taxa_taxon_db}
\alias{taxon_db<-.taxa_taxon_id}
\alias{taxon_db.taxa_taxon_id}
\alias{taxon_db<-.taxa_taxon_name}
\alias{taxon_db.taxa_taxon_name}
\alias{taxon_db<-.taxa_taxon_rank}
\alias{taxon_db.taxa_taxon_rank}
\title{Taxon database class}
\usage{
taxon_db(...)

\method{taxon_db}{default}(db = character(), ...)

taxon_db(x) <- value

\method{is.na}{taxa_taxon_db}(x)

\method{taxon_db}{taxa_taxon_id}(x) <- value

\method{taxon_db}{taxa_taxon_id}(db = character(), ...)

\method{taxon_db}{taxa_taxon_name}(x) <- value

\method{taxon_db}{taxa_taxon_name}(x, ...)

\method{taxon_db}{taxa_taxon_rank}(x) <- value

\method{taxon_db}{taxa_taxon_rank}(x, ...)
}
\arguments{
\item{...}{Used to pass arguments to methods and allow methods to used additional arguments.}

\item{db}{Zero or more taxonomic database names. Should be a name contained in
\link{db_ref}. Inputs will be transformed to a \code{character} vector if possible.}

\item{x}{An object to set taxon databases for.}

\item{value}{The taxon databases to use. Inputs will be coerced into a \code{\link[=taxon_db]{taxon_db()}} vector.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_db}
}
\description{
\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")}
Used to store the names of taxon databases defined in \code{db_ref}.
}
\examples{

# Making new objects
x <- taxon_db(c('ncbi', 'ncbi', 'itis'))
x

# Manipulating objects
as.character(x)
x[2:3]
x[x == 'itis'] <- 'gbif'
x
y = tibble::tibble(x = x)
y
y = data.frame(x = x)
y
y$x

# Trying to use an invalid database generates an error
# x <- taxon_db(c('ncbi', 'ncbi', 'my_custom_db'))
# x[x == 'itis'] <- 'my_custom_db'

# Listing known databases and their properties
db_ref$get()

# Adding and using a new database
db_ref$set(name = 'my_custom_db', desc = 'I just made this up')
db_ref$get()
x <- taxon_db(c('ncbi', 'ncbi', 'my_custom_db'))

}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}}, \code{\link{taxmap}},
  \code{\link{taxon_id}}, \code{\link{taxon_name}},
  \code{\link{taxon_rank}}, \code{\link{taxonomy}},
  \code{\link{taxon}}
}
\concept{classes}
