% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id.R
\name{taxon_id}
\alias{taxon_id}
\title{TaxonId class}
\usage{
taxon_id(id, database = NULL)
}
\arguments{
\item{id}{(character) a taxonomic id required}

\item{database}{(character) database class object, optional}
}
\value{
An \code{R6Class} object of class \code{TaxonId}
}
\description{
TaxonId class
}
\section{Methods}{

\itemize{
 \item print - print method
}
}
\examples{
(x <- taxon_id(12345))
x$id
x$database

(x <- taxon_id(
  12345,
  database_list$ncbi
))
x$id
x$database
}

