% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id.R
\name{taxon_id}
\alias{taxon_id}
\alias{names.taxa_taxon_id}
\alias{names<-.taxa_taxon_id}
\alias{is.na.taxa_taxon_id}
\title{Taxon ID class}
\usage{
taxon_id(id = character(), db = NA, .names = NULL)

\method{names}{taxa_taxon_id}(x)

\method{names}{taxa_taxon_id}(x) <- value

\method{is.na}{taxa_taxon_id}(x)
}
\arguments{
\item{id}{Zero or more taxonomic ids. Inputs will be transformed to a \code{character} vector if
possible.}

\item{db}{The name(s) of the database(s) associated with the IDs. If not \code{NA} (the default), the
input must consist of names of databases in \link{database_ref}. The length must be 0, 1, or equal
to the number of IDs.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_id}
}
\description{
\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")}
The function used to create \code{taxon_id} objects and access \code{taxon_id} objects in other classes
that contain them, such as \code{\link[=taxon]{taxon()}}, Used to store taxon IDs, either arbitrary or from a
taxonomy database. This is typically used to store taxon IDs in \code{\link[=taxon]{taxon()}} objects.
}
\examples{

# Making new objects
x <- taxon_id(c('A', 'B', 'C'))
x <- taxon_id(c(1232, 3232, 1242))
x <- taxon_id(c('9606', '1386', '4890', '4345'), db = 'ncbi')
x <- taxon_id(c('9606', '1386', '4890', '4345'),
              db = c('ncbi', 'ncbi', 'itis', 'itis'))

# Manipulating objects
as.character(x)
as.data.frame(x)
as_tibble(x)
x[2:3]
x[2:3] <- 'ABC'
names(x) <- c('a', 'b', 'c', 'd')
tax_db(x) <- 'nbn'

# Using as columns in tables
tibble::tibble(x = x, y = 1:4)
data.frame(x = x, y = 1:4)

# Trying to use an invalid ID with a specified database causes an error
#taxon_id('NOLETTERS', db = 'ncbi')

}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}},
  \code{\link{tax_date<-.taxa_taxon_authority}},
  \code{\link{taxmap}}, \code{\link{taxon_authority}},
  \code{\link{taxon_name}}, \code{\link{taxon_rank}},
  \code{\link{taxonomy}}, \code{\link{taxon}}
}
\concept{classes}
