% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{taxon_ids}
\alias{taxon_ids}
\title{Get taxon IDs}
\arguments{
\item{obj}{The \code{taxonomy} or \code{taxmap} object.}
}
\description{
Return the taxon IDs in a \code{\link[=taxonomy]{taxonomy()}} or \code{\link[=taxmap]{taxmap()}} object.
They are in the order they appear in the edge list.
\preformatted{
obj$taxon_ids()
taxon_ids(obj)}
}
\seealso{
Other taxonomy data functions: \code{\link{id_classifications}},
  \code{\link{is_branch}}, \code{\link{is_leaf}},
  \code{\link{is_root}}, \code{\link{is_stem}},
  \code{\link{n_subtaxa_1}}, \code{\link{n_subtaxa}},
  \code{\link{n_supertaxa}},
  \code{\link{name_classifications}},
  \code{\link{taxon_indexes}}, \code{\link{taxon_names}},
  \code{\link{taxon_ranks}}
}
