% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R, R/taxon_info.R
\name{taxon_info<-.taxa_taxon}
\alias{taxon_info<-.taxa_taxon}
\alias{taxon_info.taxa_taxon}
\alias{taxon_info}
\alias{taxon_info<-}
\title{Set and get taxon info}
\usage{
\method{taxon_info}{taxa_taxon}(x, key = NULL) <- value

\method{taxon_info}{taxa_taxon}(x, key = NULL, ...)

taxon_info(...)

taxon_info(x, key = NULL) <- value
}
\arguments{
\item{x}{An object with taxa.}

\item{key}{The name of the item to get from the info for each taxon. If the element with the name
specified is present, \code{NULL} will be returned/set for those taxa. If a key is not supplied
(default), then all info for each taxon will be returned/set.}

\item{value}{The taxa to set. Inputs will be coerced into a list of uniquely named list.}

\item{...}{Used to pass arguments to methods and allow methods to used additional arguments.}
}
\description{
Set and get taxon info in \code{\link[=taxon]{taxon()}} objects.
}
