% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_name.R
\name{taxon_name}
\alias{taxon_name}
\title{Taxon name class}
\usage{
taxon_name(name, database = NULL)
}
\arguments{
\item{name}{(character) a taxonomic name. required}

\item{database}{(character) database class object, optional}
}
\value{
An \code{R6Class} object of class \code{TaxonName}
}
\description{
Used to store the name of taxa. This is typically used to
store where taxon names in \code{\link[=taxon]{taxon()}} objects.
}
\examples{
(poa <- taxon_name("Poa"))
(undef <- taxon_name("undefined"))
(sp1 <- taxon_name("species 1"))
(poa_annua <- taxon_name("Poa annua"))
(x <- taxon_name("Poa annua L."))

x$name
x$database

(x <- taxon_name(
  "Poa annua",
  database_list$ncbi
))
x$rank
x$database
}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}}, \code{\link{taxa_}},
  \code{\link{taxmap}}, \code{\link{taxon_database}},
  \code{\link{taxon_id}}, \code{\link{taxon_rank}},
  \code{\link{taxonomy}}, \code{\link{taxon}}
}
