% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_name.R
\name{taxon_name}
\alias{taxon_name}
\title{Taxon name class}
\usage{
taxon_name(name = character(0), rank = NA, id = NA, auth = NA)
}
\arguments{
\item{name}{The names of taxa. Inputs with be coerced into a \link{taxon_name} vector if anything else
is given.}

\item{rank}{The ranks of taxa. Inputs with be coerced into a \link{taxon_rank} vector if anything else
is given.}

\item{id}{The ids of taxa. These should be unique identifier and are usually associated with a
database. Inputs with be coerced into a \link{taxon_id} vector if anything else is given.}

\item{auth}{The authority of the taxon. Inputs with be coerced into a \link{character} vector if
anything else is given.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_name}
}
\description{
\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")}
Used to store information about taxa, such as names, ranks, and IDs.
This should be used when storing information about taxa when things like synonyms or conflicting taxonomies are not needed.
To support the possiblity of synonyms or encode confliciting taxonomic opinions, use the \code{\link[=taxon]{taxon()}}, \code{\link[=taxonomy]{taxonomy()}}, \code{\link[=hierarchy]{hierarchy()}}, or \code{\link[=taxmap]{taxmap()}} classes.
For more information on what each class is designed for, see the \link{concepts} section of the help pages.
}
\examples{

# Create taxon name vector
x <- taxon_name(c('A', 'B', 'C'))
x <- taxon_name(name = c('Homo sapiens', 'Bacillus', 'Ascomycota', 'Ericaceae'),
                rank = c('species', 'genus', 'phylum', 'family'),
                id = taxon_id(c('9606', '1386', '4890', '4345'), db = 'ncbi'),
                auth = c('Linnaeus, 1758', 'Cohn 1872', NA, 'Juss., 1789'))

# Get parts of the taxon name vector
tax_name(x)
tax_rank(x)
tax_id(x)
tax_auth(x)

# Set parts of the taxon name vector
tax_name(x) <- tolower(tax_name(x))
tax_rank(x)[1] <- NA
tax_db(taxon_id(x)) <- 'itis'
tax_auth(x) <- NA

# Manipulate taxon name vectors
x[1:3]
x[tax_rank(x) > 'family']
c(x, x)

}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}},
  \code{\link{levels<-.taxa_taxon}},
  \code{\link{tax_auth_date<-}}, \code{\link{taxmap}},
  \code{\link{taxon_authority}}, \code{\link{taxon_id}},
  \code{\link{taxonomy}}
}
\concept{classes}
