% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--docs.R
\name{taxon_names}
\alias{taxon_names}
\title{Get taxon names}
\arguments{
\item{obj}{The \code{taxonomy} or \code{taxmap} object.}
}
\description{
Return the taxon names in a \code{\link[=taxmap]{taxmap()}} object.
They are in the order they appear in the edge list.
}
\details{
\preformatted{
obj$taxon_names()
taxon_names(obj, ...)}
}
