% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank.R
\name{taxon_rank}
\alias{taxon_rank}
\title{Taxon rank class}
\usage{
taxon_rank(rank = character(), db = NA, levels = NULL,
  guess_order = TRUE)
}
\arguments{
\item{rank}{Zero or more taxonomic rank names. Inputs will be transformed to
a \code{character} vector.}

\item{db}{The name(s) of the database(s) associated with the IDs. If not \code{NA}
(the default), the input must consist of names of databases in
\link{database_list}. The length must be 0, 1, or equal to the number of IDs.}

\item{levels}{A named numeric vector indicating the names and orders of
possible taxonomic ranks. Higher numbers indicate for fine-scale groupings.
Ranks of unknown order can be indicated with \code{NA} instead of a number.}

\item{guess_order}{If \code{TRUE} and no rank order is given using numbers, try to guess order based on rank names.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_rank}
}
\description{
Used to store taxon ranks, possibly assocaited with a taxonomy database. This is typically used to
store taxon ranks in \code{\link[=taxon]{taxon()}} objects.
}
\examples{
(x <- taxon_rank(12345))
x$rank
x$database

(x <- taxon_rank(
  12345,
  database_list$ncbi
))
x$rank
x$database

# a null taxon_name object
taxon_name(NULL)

}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}}, \code{\link{taxa}},
  \code{\link{taxmap}}, \code{\link{taxon_auth}},
  \code{\link{taxon_db_def}}, \code{\link{taxon_db}},
  \code{\link{taxon_id}}, \code{\link{taxon_info}},
  \code{\link{taxon_name}}, \code{\link{taxon_rank_level}},
  \code{\link{taxonomy}}, \code{\link{taxon}}
}
\concept{classes}
