% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_rank.R
\name{taxon_rank}
\alias{taxon_rank}
\title{Taxon ID class}
\usage{
taxon_rank(rank = character(), db = NA)
}
\arguments{
\item{rank}{Zero or more taxonomic ids. Inputs will be transformed to a \code{character} vector.}

\item{db}{The name(s) of the database(s) associated with the IDs. If not \code{NA} (the
default), the input must consist of names of databases in \link{database_list}. The length must be
0, 1, or equal to the number of IDs.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_rank}
}
\description{
Used to store taxon IDs, either arbitrary or from a taxonomy database. This is typically used to
store taxon IDs in \code{\link[=taxon]{taxon()}} objects.
}
\examples{
(x <- taxon_rank(12345))
x$rank
x$database

(x <- taxon_rank(
  12345,
  database_list$ncbi
))
x$rank
x$database

# a null taxon_name object
taxon_name(NULL)

}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}}, \code{\link{taxa}},
  \code{\link{taxmap}}, \code{\link{taxon_db_def}},
  \code{\link{taxon_db}}, \code{\link{taxon_id}},
  \code{\link{taxon_name}}, \code{\link{taxonomy}},
  \code{\link{taxon}}
}
\concept{classes}
