% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R, R/taxon_rank.R
\name{levels<-.taxa_taxon}
\alias{levels<-.taxa_taxon}
\alias{levels.taxa_taxon}
\alias{taxon_rank<-.taxa_taxon}
\alias{taxon_rank.taxa_taxon}
\alias{taxon_rank}
\alias{taxon_rank.default}
\alias{taxon_rank<-}
\alias{levels<-.taxa_taxon_rank}
\alias{levels.taxa_taxon_rank}
\alias{is.na.taxa_taxon_rank}
\title{Taxon rank class}
\usage{
\method{levels}{taxa_taxon}(x) <- value

\method{levels}{taxa_taxon}(x)

\method{taxon_rank}{taxa_taxon}(x) <- value

\method{taxon_rank}{taxa_taxon}(x, ...)

taxon_rank(...)

\method{taxon_rank}{default}(rank = character(), db = NA,
  levels = NULL, guess_order = TRUE, ...)

taxon_rank(x) <- value

\method{levels}{taxa_taxon_rank}(x) <- value

\method{levels}{taxa_taxon_rank}(x)

\method{is.na}{taxa_taxon_rank}(x)
}
\arguments{
\item{x}{An object with taxon ranks.}

\item{value}{The ranks to set. Inputs will be coerced into a \code{\link[=taxon_rank]{taxon_rank()}} vector.}

\item{...}{Used to pass arguments to methods and allow methods to use additional arguments.}

\item{rank}{Zero or more taxonomic rank names. Inputs will be transformed to a \code{character}
vector.}

\item{db}{The name(s) of the database(s) associated with the IDs. If not \code{NA} (the default), the
input must consist of names of databases in \link{database_ref}. The length must be 0, 1, or equal
to the number of IDs.}

\item{levels}{A named numeric vector indicating the names and orders of possible taxonomic ranks.
Higher numbers indicate for fine-scale groupings. Ranks of unknown order can be indicated with
\code{NA} instead of a number.}

\item{guess_order}{If \code{TRUE} and no rank order is given using numbers, try to guess order based
on rank names.}
}
\value{
An \code{S3} object of class \code{taxa_taxon_rank}
}
\description{
\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")}
Used to store taxon ranks, possibly assocaited with a taxonomy database. This is typically used to
store taxon ranks in \code{\link[=taxon]{taxon()}} objects.
}
\examples{

# Making new objects
x <- taxon_rank(c('species', 'species', 'phylum', 'family'))
x <- taxon_rank(c('species', 'species', 'phylum', 'family'),
                db = c('ncbi', 'ncbi', 'itis', 'itis'))

# Specifiying level order
taxon_rank(c('A', 'B', 'C', 'D', 'A', 'D', 'D'),
           levels = c('D', 'C', 'B', 'A'))
taxon_rank(c('A', 'B', 'C', 'D', 'A', 'D', 'D'),
           levels = c(D = NA, A = 10, B = 20, C = 30))

# Manipulating objects
as.character(x)
as.factor(x)
as.ordered(x)
x[2:3]
x[x > 'family'] <- taxon_rank('unknown')
x[1] <- taxon_rank('order')
x[4] <- taxon_rank('order', db = 'ncbi')

# Using as columns in tables
tibble::tibble(x = x, y = 1:4)
data.frame(x = x, y = 1:4)

# Trying to add an unknown level causes an error
#x[2] <- 'unknown'

}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}}, \code{\link{taxmap}},
  \code{\link{taxon_db_def}}, \code{\link{taxon_db}},
  \code{\link{taxon_id<-.taxa_taxon}},
  \code{\link{taxon_name<-.taxa_taxon}},
  \code{\link{taxon_rank_level}}, \code{\link{taxonomy}},
  \code{\link{taxon}}
}
\concept{classes}
