% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{taxon_rank}
\alias{taxon_rank}
\title{TaxonRank class}
\usage{
taxon_rank(name, database = NULL)
}
\arguments{
\item{name}{(character) rank name. required}

\item{database}{(character) database class object, optional}
}
\value{
An \code{R6Class} object of class \code{TaxonRank}
}
\description{
TaxonRank class
}
\examples{
taxon_rank("species")
taxon_rank("genus")
taxon_rank("kingdom")

(x <- taxon_rank(
  "species",
  database_list$ncbi
))
x$rank
x$database
}
