% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy--class.R, R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\alias{taxonomy.default}
\title{Taxonomy class}
\usage{
taxonomy(...)

taxonomy(...)

\method{taxonomy}{default}(taxa = taxon(), supertaxa = integer(), ...)
}
\arguments{
\item{...}{Any number of object of class \code{\link[=hierarchy]{hierarchy()}} or character
vectors.}

\item{taxa}{A \code{\link[=taxon]{taxon()}} vector.}

\item{supertaxa}{The indexes of \code{taxa} for each taxon's supertaxon.}

\item{.list}{An alternate to the \code{...} input. Any number of object of class
\code{\link[=hierarchy]{hierarchy()}} or character vectors in a list. Cannot be used with \code{...}.}

\item{...}{Used to pass arguments to methods and allow methods to used additional arguments.}
}
\value{
An \code{R6Class} object of class \code{Taxonomy}

An \code{S3} object of class \code{taxa_taxon}
}
\description{
Stores a taxonomy composed of \code{\link[=taxon]{taxon()}} objects organized in a tree
structure. This differs from the \code{\link[=hierarchies]{hierarchies()}} class in how the \code{\link[=taxon]{taxon()}}
objects are stored. Unlike \code{\link[=hierarchies]{hierarchies()}}, each taxon is only stored once
and the relationships between taxa are stored in an \href{https://en.wikipedia.org/wiki/Adjacency_list}{edge list}.

\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")}
Used to store information about a set of taxa forming a taxonomic tree.
}
\examples{

# Making a taxonomy object with vectors
taxonomy(c("mammalia", "felidae", "panthera", "tigris"),
         c("mammalia", "felidae", "panthera", "leo"),
         c("mammalia", "felidae", "felis", "catus"))

# Making a taxonomy object from scratch
#   Note: This information would usually come from a parsing function.
#         This is just for demonstration.
x <- taxon(
  name = taxon_name("Notoryctidae"),
  rank = taxon_rank("family"),
  id = taxon_id(4479)
)
y <- taxon(
  name = taxon_name("Notoryctes"),
  rank = taxon_rank("genus"),
  id = taxon_id(4544)
)
z <- taxon(
  name = taxon_name("Notoryctes typhlops"),
  rank = taxon_rank("species"),
  id = taxon_id(93036)
)

a <- taxon(
  name = taxon_name("Mammalia"),
  rank = taxon_rank("class"),
  id = taxon_id(9681)
)
b <- taxon(
  name = taxon_name("Felidae"),
  rank = taxon_rank("family"),
  id = taxon_id(9681)
)

cc <- taxon(
  name = taxon_name("Puma"),
  rank = taxon_rank("genus"),
  id = taxon_id(146712)
)
d <- taxon(
  name = taxon_name("Puma concolor"),
  rank = taxon_rank("species"),
  id = taxon_id(9696)
)

m <- taxon(
  name = taxon_name("Panthera"),
  rank = taxon_rank("genus"),
  id = taxon_id(146712)
)
n <- taxon(
  name = taxon_name("Panthera tigris"),
  rank = taxon_rank("species"),
  id = taxon_id(9696)
)

(hier1 <- hierarchy(z, y, x, a))
(hier2 <- hierarchy(cc, b, a, d))
(hier3 <- hierarchy(n, m, b, a))

(hrs <- hierarchies(hier1, hier2, hier3))

ex_taxonomy <- taxonomy(hier1, hier2, hier3)


}
\seealso{
Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}},
  \code{\link{levels<-.taxa_taxon}},
  \code{\link{tax_auth_date<-}}, \code{\link{taxmap}},
  \code{\link{taxon_authority}}, \code{\link{taxon_id}},
  \code{\link{taxon_name}}

Other classes: \code{\link{hierarchies}},
  \code{\link{hierarchy}},
  \code{\link{levels<-.taxa_taxon}},
  \code{\link{tax_auth_date<-}}, \code{\link{taxmap}},
  \code{\link{taxon_authority}}, \code{\link{taxon_id}},
  \code{\link{taxon_name}}
}
\concept{classes}
