% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\alias{names.taxa_taxonomy}
\alias{names<-.taxa_taxonomy}
\title{Taxonomy class}
\usage{
taxonomy(taxa = taxon(), supertaxa = integer())

\method{names}{taxa_taxonomy}(x)

\method{names}{taxa_taxonomy}(x) <- value
}
\arguments{
\item{taxa}{A \code{\link[=taxon]{taxon()}} vector.}

\item{supertaxa}{The indexes of \code{taxa} for each taxon's supertaxon.}
}
\value{
An \code{S3} object of class \code{taxa_taxon}
}
\description{
\Sexpr[results=rd, stage=render]{taxa:::lifecycle("experimental")}
Used to store information about a set of taxa forming a taxonomic tree.
}
\examples{

x <- taxonomy(c('Carnivora', 'Felidae', 'Panthera', 'Panthera leo',
                'Panthera tigris', 'Ursidae', 'Ursus', 'Ursus arctos'),
              supertaxa = c(NA, 1, 2, 3, 3, 1, 6, 7))



x <- taxonomy(taxon(name = c('Carnivora', 'Felidae', 'Panthera', 'Panthera leo',
                             'Panthera tigris', 'Ursidae', 'Ursus', 'Ursus arctos'),
                    rank = c('order', 'family', 'genus', 'species',
                             'species', 'family', 'genus', 'species'),
                    id = taxon_id(c('33554', '9681', '9688', '9689',
                                    '9694', '9632', '9639', '9644'),
                                  db = 'ncbi'),
                    auth = c('Bowdich, 1821', 'Fischer de Waldheim, 1817', 'Oken, 1816', 'L., 1758',
                             'L., 1758', 'Fischer de Waldheim, 1817', 'L., 1758', 'L., 1758')),
              supertaxa = c(NA, 1, 2, 3, 3, 1, 6, 7))

# Get parts of the taxonomy vector
tax_name(x)
tax_rank(x)
tax_id(x)
tax_db(x)
tax_auth(x)
tax_author(x)
tax_date(x)
tax_cite(x)

# Set parts of the taxonomy vector
tax_name(x) <- tolower(tax_name(x))
tax_rank(x)[1] <- NA
tax_id(x) <- '9999'
tax_db(x) <- 'itis'
tax_auth(x) <- NA
tax_author(x)[2:3] <- c('Joe', 'Billy')
tax_date(x) <- c('1999', '2013', '1796', '1899')
tax_cite(x)[1] <- 'Linnaeus, C. (1771). Mantissa plantarum altera generum.'

}
\seealso{
Other classes: \code{\link{taxon_authority}},
  \code{\link{taxon_db}}, \code{\link{taxon_id}},
  \code{\link{taxon_name}}, \code{\link{taxon_rank}},
  \code{\link{taxon}}
}
\concept{classes}
