% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxmap--internal.R
\name{validate_taxmap_data}
\alias{validate_taxmap_data}
\title{Convert \code{data} input for Taxamp}
\usage{
validate_taxmap_data(data, input_ids)
}
\arguments{
\item{data}{The \code{data} variable passed to the \code{Taxmap} constructor}

\item{input_ids}{}
}
\value{
A \code{data} varaible with the right format
}
\description{
Make sure \code{data} is in the right format and complain if it is not.
Then, add a \code{taxon_id} column to data with the same length as the input
}
\keyword{internal}
